Split focus.rs into focused submodules (2,946 lines → ~1,250 max) (#491)
Split src/focus.rs into src/focus/ directory with 7 submodules:
- mod.rs (47 lines): Public API, re-exports
- layers.rs (167 lines): Network layer computation via BFS
- allocation.rs (215 lines): Budget allocation strategies
- gradient.rs (582 lines): Gradient flow analysis
- impact.rs (653 lines): Impact calculation
- ranking.rs (1,248 lines): Neuron ranking, record providers
- tests.rs (71 lines): Unit tests for internal components

No changes to the public API — all existing imports continue to work.
All 463 unit tests and ~97 integration test files pass unchanged.
