--- src/focus.rs
@@ -1,2946 +1,0 @@
-use crate::analysis::{check_memory_for_parquet, verbose_enabled};
-use crate::discovery_history::DiscoveryHistory;
-use crate::intern::NeuronIndex;
-use crate::parquet_format::{read_all_records_grouped_by_neuron, read_records_from_parquet};
-use crate::types::DiscoverRecord;
-use crate::{
-    CoordinatedStructuralCandidateJson, CoordinatedStructuralOpJson, CreatureJson, NeuronJson,
-    SynapseJson,
-};
-use anyhow::{Context, Result, anyhow};
-use rayon::prelude::*;
-
-use std::collections::{HashMap, HashSet, VecDeque};
-use std::sync::{Arc, Mutex};
-use std::time::Instant;
-
-/// Statistics for selection-based neurons (MINIMUM, MAXIMUM, IF).
-/// Maps (from_uuid, to_uuid) -> win probability (0.0 to 1.0).
-/// For MIN/MAX: probability this synapse provides the min/max value.
-/// For IF: probability this synapse's branch is taken (condition always 1.0).
-pub type SelectionStats = HashMap<(String, String), f32>;
-
-/// A synapse key paired with its weighted activation contribution.
-/// Used internally for tracking which synapse wins in MIN/MAX calculations.
-type SynapseContribution = ((String, String), f32);
-
-/// Map from observation index to list of synapse contributions for that observation.
-/// Used to determine which synapse wins (has min/max value) for each observation.
-type ObservationContributions = HashMap<u32, Vec<SynapseContribution>>;
-
-// =============================================================================
-// Hierarchical Focus Selection (Issue #222)
-// =============================================================================
-
-/// Represents a layer of neurons at a specific depth in the network.
-///
-/// Neurons are grouped by their distance from inputs (measured in synapse hops).
-/// This enables hierarchical focus selection that guarantees coverage across
-/// all network depths.
-#[derive(Debug, Clone)]
-pub struct NeuronLayer {
-    /// Distance from inputs (0 = directly connected to inputs)
-    pub depth: usize,
-    /// Neurons at this depth (references to creature's neuron list)
-    pub neurons: Vec<NeuronInfo>,
-}
-
-/// Basic neuron information for layer assignment.
-#[derive(Debug, Clone)]
-pub struct NeuronInfo {
-    pub uuid: String,
-    pub neuron_type: String,
-}
-
-/// Strategy for allocating focus budget across layers.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AllocationStrategy {
-    /// Equal allocation per layer (total / num_layers)
-    Equal,
-    /// Proportional to layer size (larger layers get more slots)
-    Proportional,
-    /// Prioritise output layers (allocate from deepest to shallowest)
-    OutputFirst,
-}
-
-/// Compute network layers by organising neurons based on their depth from inputs.
-///
-/// Depth is computed using BFS from input neurons:
-/// - Depth 0: Input neurons (not included in returned layers as they're not selectable)
-/// - Depth 1: Neurons directly connected to inputs
-/// - Depth N: Neurons at N hops from any input
-///
-/// # Algorithm
-///
-/// Uses BFS to compute the MAXIMUM depth for each neuron. When a neuron has
-/// multiple incoming paths, we use the longest path to determine its layer.
-/// This ensures output-adjacent neurons are in deeper layers even if they
-/// have short paths from some inputs.
-///
-/// # Handling edge cases
-///
-/// - **Cycles**: Detected and handled by tracking visited nodes. A node in a cycle
-///   gets its depth from the first path that reaches it.
-/// - **Disconnected components**: Neurons not reachable from inputs are assigned
-///   to a special "unreachable" layer (depth = usize::MAX, sorted last).
-///
-/// # Arguments
-/// * `creature` - The creature to analyse
-///
-/// # Returns
-/// Vector of NeuronLayers sorted by depth (shallowest first)
-pub fn compute_network_layers(creature: &CreatureJson) -> Vec<NeuronLayer> {
-    // Issue #210: Use NeuronIndex for memory-efficient BFS traversal.
-    // Instead of cloning UUID strings for the adjacency map and queue,
-    // we use u32 indices (4 bytes vs ~36+ bytes per String).
-    let mut neuron_index = NeuronIndex::with_capacity(creature.neurons.len() + creature.input);
-
-    // Pre-intern all neuron UUIDs
-    for neuron in &creature.neurons {
-        neuron_index.intern(&neuron.uuid);
-    }
-
-    // Build neuron UUID set using indices for efficient lookup
-    let neuron_indices: HashSet<u32> = creature
-        .neurons
-        .iter()
-        .map(|n| neuron_index.get_index(&n.uuid).unwrap())
-        .collect();
-
-    // Build forward adjacency using interned indices
-    let mut forward_adjacency: HashMap<u32, Vec<u32>> = HashMap::new();
-    for synapse in &creature.synapses {
-        let from_idx = neuron_index.intern(&synapse.from_uuid);
-        let to_idx = neuron_index.intern(&synapse.to_uuid);
-        forward_adjacency.entry(from_idx).or_default().push(to_idx);
-    }
-
-    // Identify input neuron indices (those not in the neuron list)
-    let input_indices: HashSet<u32> = creature
-        .synapses
-        .iter()
-        .map(|s| neuron_index.get_index(&s.from_uuid).unwrap())
-        .filter(|idx| !neuron_indices.contains(idx))
-        .collect();
-
-    // BFS to compute depth from inputs using indices
-    let mut depths: HashMap<u32, usize> = HashMap::new();
-    let mut queue: VecDeque<u32> = VecDeque::new();
-
-    // Maximum depth to prevent infinite loops with cycles
-    let max_depth = creature.neurons.len() + creature.input + 1;
-
-    // Initialise: all input indices have depth 0
-    for &idx in &input_indices {
-        depths.insert(idx, 0);
-        queue.push_back(idx);
-    }
-
-    // BFS traversal with cycle protection
-    while let Some(idx) = queue.pop_front() {
-        let current_depth = depths[&idx];
-
-        // Stop propagating if we've exceeded max depth (cycle detection)
-        if current_depth >= max_depth {
-            continue;
-        }
-
-        if let Some(targets) = forward_adjacency.get(&idx) {
-            for &to_idx in targets {
-                let new_depth = current_depth + 1;
-
-                // Only update if we haven't seen this node or found a longer path
-                // BUT cap at max_depth to handle cycles
-                if new_depth <= max_depth {
-                    let should_update = match depths.get(&to_idx) {
-                        Some(&existing) => new_depth > existing && new_depth <= max_depth,
-                        None => true,
-                    };
-
-                    if should_update {
-                        depths.insert(to_idx, new_depth);
-                        queue.push_back(to_idx);
-                    }
-                }
-            }
-        }
-    }
-
-    // Handle disconnected neurons (not reachable from any input)
-    for neuron in &creature.neurons {
-        let idx = neuron_index.get_index(&neuron.uuid).unwrap();
-        depths.entry(idx).or_insert(usize::MAX);
-    }
-
-    // Group neurons by depth
-    let mut layers_map: HashMap<usize, Vec<NeuronInfo>> = HashMap::new();
-    for neuron in &creature.neurons {
-        // Skip input and constant neurons (not selectable)
-        if neuron.neuron_type == "input" || neuron.neuron_type == "constant" {
-            continue;
-        }
-
-        let idx = neuron_index.get_index(&neuron.uuid).unwrap();
-        let depth = depths.get(&idx).copied().unwrap_or(usize::MAX);
-        layers_map.entry(depth).or_default().push(NeuronInfo {
-            uuid: neuron.uuid.clone(),
-            neuron_type: neuron.neuron_type.clone(),
-        });
-    }
-
-    // Convert to sorted vector
-    let mut layers: Vec<NeuronLayer> = layers_map
-        .into_iter()
-        .map(|(depth, neurons)| NeuronLayer { depth, neurons })
-        .collect();
-
-    // Sort by depth (shallowest first, unreachable last)
-    layers.sort_by_key(|l| l.depth);
-
-    layers
-}
-
-/// Allocate focus budget across layers based on the specified strategy.
-///
-/// # Arguments
-/// * `total` - Total focus budget to allocate
-/// * `layers` - Network layers (from compute_network_layers)
-/// * `strategy` - Allocation strategy to use
-///
-/// # Returns
-/// Vector of allocations, one per layer, in the same order as `layers`
-fn allocate_focus_budget(
-    total: usize,
-    layers: &[NeuronLayer],
-    strategy: AllocationStrategy,
-) -> Vec<usize> {
-    if layers.is_empty() {
-        return Vec::new();
-    }
-
-    match strategy {
-        AllocationStrategy::Equal => {
-            // Divide equally among layers
-            let per_layer = total / layers.len();
-            let remainder = total % layers.len();
-            let mut alloc = vec![per_layer; layers.len()];
-
-            // Distribute remainder to last layers (closer to output)
-            for i in 0..remainder {
-                let idx = layers.len() - 1 - i;
-                alloc[idx] += 1;
-            }
-            alloc
-        }
-
-        AllocationStrategy::Proportional => {
-            // Allocate proportionally to layer size
-            let total_neurons: usize = layers.iter().map(|l| l.neurons.len()).sum();
-            if total_neurons == 0 {
-                return vec![0; layers.len()];
-            }
-
-            let mut alloc: Vec<usize> = layers
-                .iter()
-                .map(|l| (total * l.neurons.len()) / total_neurons)
-                .collect();
-
-            // Distribute any remainder
-            let allocated: usize = alloc.iter().sum();
-            let remainder = total.saturating_sub(allocated);
-            for i in 0..remainder {
-                let idx = layers.len() - 1 - (i % layers.len());
-                alloc[idx] += 1;
-            }
-            alloc
-        }
-
-        AllocationStrategy::OutputFirst => {
-            // Prioritise layers closest to output (deepest first)
-            let mut alloc = vec![0usize; layers.len()];
-            let mut remaining = total;
-
-            // Process from deepest to shallowest
-            for i in (0..layers.len()).rev() {
-                if remaining == 0 {
-                    break;
-                }
-                // Take at most half of remaining budget for this layer,
-                // but cap at layer size
-                let take = (remaining / 2).max(1).min(layers[i].neurons.len());
-                alloc[i] = take;
-                remaining = remaining.saturating_sub(take);
-            }
-
-            // If we still have budget, distribute to layers that can take more
-            for i in (0..layers.len()).rev() {
-                if remaining == 0 {
-                    break;
-                }
-                let can_take = layers[i].neurons.len().saturating_sub(alloc[i]);
-                let take = can_take.min(remaining);
-                alloc[i] += take;
-                remaining = remaining.saturating_sub(take);
-            }
-
-            alloc
-        }
-    }
-}
-
-/// Select focus neurons hierarchically, ensuring coverage across all network layers.
-///
-/// This function implements hierarchical focus selection for large creatures,
-/// addressing the issue where flat selection tends to over-represent certain
-/// layers (typically those with the highest raw error scores).
-///
-/// # Algorithm
-///
-/// 1. Allocate focus budget across layers using the specified strategy
-/// 2. Within each layer, select the top-scoring neurons up to the allocation
-/// 3. If a layer doesn't have enough neurons, redistribute its unused slots
-///
-/// # Arguments
-/// * `creature` - The creature (used for validation)
-/// * `layers` - Pre-computed network layers from `compute_network_layers`
-/// * `max_focus` - Maximum number of neurons to select
-/// * `strategy` - How to allocate budget across layers
-/// * `scores` - Map from neuron UUID to score (higher = more likely to select)
-///
-/// # Returns
-/// Vector of selected neuron UUIDs
-pub fn hierarchical_focus_selection(
-    _creature: &CreatureJson,
-    layers: &[NeuronLayer],
-    max_focus: usize,
-    strategy: AllocationStrategy,
-    scores: &HashMap<String, f32>,
-) -> Vec<String> {
-    if layers.is_empty() || max_focus == 0 {
-        return Vec::new();
-    }
-
-    // Allocate budget across layers
-    let allocations = allocate_focus_budget(max_focus, layers, strategy);
-
-    let mut selected: Vec<String> = Vec::with_capacity(max_focus);
-    let mut unused_slots = 0usize;
-
-    // Select from each layer
-    for (i, layer) in layers.iter().enumerate() {
-        let allocation = allocations[i] + unused_slots;
-        unused_slots = 0;
-
-        if allocation == 0 {
-            continue;
-        }
-
-        // Sort neurons in this layer by score (descending)
-        let mut layer_neurons: Vec<(&NeuronInfo, f32)> = layer
-            .neurons
-            .iter()
-            .map(|n| (n, scores.get(&n.uuid).copied().unwrap_or(0.0)))
-            .collect();
-
-        layer_neurons.sort_by(|a, b| b.1.total_cmp(&a.1).then_with(|| a.0.uuid.cmp(&b.0.uuid)));
-
-        // Select top neurons from this layer
-        let to_select = allocation.min(layer_neurons.len());
-        for (neuron, _score) in layer_neurons.into_iter().take(to_select) {
-            selected.push(neuron.uuid.clone());
-        }
-
-        // Track unused slots for redistribution
-        unused_slots = allocation.saturating_sub(to_select);
-    }
-
-    // If we still have unused slots and haven't reached max_focus,
-    // go back and select more from layers that have neurons left
-    if selected.len() < max_focus && unused_slots > 0 {
-        let already_selected: HashSet<_> = selected.iter().cloned().collect();
-
-        for layer in layers.iter().rev() {
-            // Output layers first (reversed order)
-            if unused_slots == 0 {
-                break;
-            }
-
-            let mut remaining: Vec<(&NeuronInfo, f32)> = layer
-                .neurons
-                .iter()
-                .filter(|n| !already_selected.contains(&n.uuid))
-                .map(|n| (n, scores.get(&n.uuid).copied().unwrap_or(0.0)))
-                .collect();
-
-            remaining.sort_by(|a, b| b.1.total_cmp(&a.1).then_with(|| a.0.uuid.cmp(&b.0.uuid)));
-
-            for (neuron, _score) in remaining {
-                if unused_slots == 0 {
-                    break;
-                }
-                selected.push(neuron.uuid.clone());
-                unused_slots -= 1;
-            }
-        }
-    }
-
-    // Ensure we don't exceed max_focus
-    selected.truncate(max_focus);
-
-    selected
-}
-
-/// Threshold for using hierarchical selection (number of selectable neurons).
-/// Below this threshold, flat selection is used for simplicity.
-/// Above this threshold, hierarchical selection provides better coverage.
-pub const HIERARCHICAL_SELECTION_THRESHOLD: usize = 100;
-
-/// Provides access to recorded discovery data without assuming an in-memory HashMap.
-/// Implementations may pre-load all records or stream them on demand with bounded caching.
-pub trait RecordProvider: Send + Sync {
-    fn get(&self, neuron_uuid: &str) -> Result<Option<Arc<Vec<DiscoverRecord>>>>;
-    fn len(&self) -> usize;
-    fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-}
-
-type RecordLoader = dyn Fn(&str, &str) -> Result<Vec<DiscoverRecord>> + Send + Sync + 'static;
-
-struct EagerRecordProvider {
-    records: HashMap<String, Arc<Vec<DiscoverRecord>>>,
-}
-
-impl EagerRecordProvider {
-    fn new(records: HashMap<String, Vec<DiscoverRecord>>) -> Self {
-        let records = records
-            .into_iter()
-            .map(|(uuid, mut recs)| {
-                recs.sort_by_key(|r| r.obs_index);
-                (uuid, Arc::new(recs))
-            })
-            .collect();
-        Self { records }
-    }
-}
-
-impl RecordProvider for EagerRecordProvider {
-    fn get(&self, neuron_uuid: &str) -> Result<Option<Arc<Vec<DiscoverRecord>>>> {
-        Ok(self.records.get(neuron_uuid).map(Arc::clone))
-    }
-
-    fn len(&self) -> usize {
-        self.records.len()
-    }
-}
-
-struct LazyRecordProvider {
-    parquet_file: String,
-    cache: Mutex<LazyCache>,
-    loader: Arc<RecordLoader>,
-}
-
-struct LazyCache {
-    entries: HashMap<String, Arc<Vec<DiscoverRecord>>>,
-    order: VecDeque<String>,
-    capacity: usize,
-}
-
-impl LazyCache {
-    fn new(capacity: usize) -> Self {
-        Self {
-            entries: HashMap::new(),
-            order: VecDeque::new(),
-            capacity,
-        }
-    }
-
-    fn insert(&mut self, key: String, value: Arc<Vec<DiscoverRecord>>) {
-        if self.entries.contains_key(&key) {
-            self.order.retain(|k| k != &key);
-        }
-
-        self.entries.insert(key.clone(), value);
-        self.order.push_back(key);
-        self.evict();
-    }
-
-    fn evict(&mut self) {
-        while self.entries.len() > self.capacity {
-            if let Some(oldest) = self.order.pop_front() {
-                self.entries.remove(&oldest);
-            }
-        }
-    }
-}
-
-impl LazyRecordProvider {
-    const DEFAULT_CACHE_CAPACITY: usize = 8;
-
-    fn new(parquet_file: &str) -> Self {
-        Self {
-            parquet_file: parquet_file.to_string(),
-            cache: Mutex::new(LazyCache::new(Self::DEFAULT_CACHE_CAPACITY)),
-            loader: Arc::new(read_records_from_parquet),
-        }
-    }
-
-    #[cfg(test)]
-    fn with_loader_for_tests(
-        parquet_file: &str,
-        capacity: usize,
-        loader: Arc<RecordLoader>,
-    ) -> Self {
-        Self {
-            parquet_file: parquet_file.to_string(),
-            cache: Mutex::new(LazyCache::new(capacity)),
-            loader,
-        }
-    }
-}
-
-impl RecordProvider for LazyRecordProvider {
-    fn get(&self, neuron_uuid: &str) -> Result<Option<Arc<Vec<DiscoverRecord>>>> {
-        {
-            let cache = self.cache.lock().expect("lazy record cache poisoned");
-            if let Some(records) = cache.entries.get(neuron_uuid) {
-                return Ok(Some(Arc::clone(records)));
-            }
-        }
-
-        let mut records = (self.loader)(&self.parquet_file, neuron_uuid).with_context(|| {
-            format!(
-                "Failed to load discovery records for {neuron_uuid} from {}",
-                self.parquet_file
-            )
-        })?;
-        if records.is_empty() {
-            return Ok(None);
-        }
-        records.sort_by_key(|r| r.obs_index);
-        let arc_records = Arc::new(records);
-
-        let mut cache = self.cache.lock().expect("lazy record cache poisoned");
-        cache.insert(neuron_uuid.to_string(), Arc::clone(&arc_records));
-        Ok(Some(arc_records))
-    }
-
-    fn len(&self) -> usize {
-        let cache = self.cache.lock().expect("lazy record cache poisoned");
-        cache.entries.len()
-    }
-}
-
-fn get_records_or_error(
-    provider: &dyn RecordProvider,
-    neuron_uuid: &str,
-) -> Result<Arc<Vec<DiscoverRecord>>> {
-    provider
-        .get(neuron_uuid)?
-        .ok_or_else(|| anyhow!("Missing discovery records for selectable neuron: {neuron_uuid}"))
-}
-
-#[derive(Debug)]
-pub struct RankedNeuron {
-    pub neuron_uuid: String,
-    pub total_error: f32,
-    /// Unclamped mean absolute error from recorded samples.
-    ///
-    /// We clamp `total_error` for focus ranking so hidden neurons with extreme raw errors do not
-    /// dominate selection purely due to scale differences. However, when proposing exploratory
-    /// ablation candidates we still want access to the true magnitude.
-    pub raw_error: f32,
-    /// Structural impact based on weight paths to output
-    pub impact: f32,
-    /// Mean absolute activation value from recorded samples
-    pub mean_activation: f32,
-    /// Activation-weighted impact = structural_impact × mean_activation
-    /// This reflects the actual contribution the neuron makes during inference
-    pub activation_weighted_impact: f32,
-    /// Issue #206: Gradient flow statistics for this neuron.
-    ///
-    /// These metrics help identify neurons with high learning potential:
-    /// - `avg_gradient_magnitude`: How much error signal can flow through
-    /// - `saturation_ratio`: % of samples in saturated activation region
-    /// - `dead_ratio`: % of samples with zero gradient (ReLU dead zones)
-    pub gradient_flow: GradientFlowStats,
-    /// Issue #204: Activation frequency (proportion of samples where neuron fires).
-    ///
-    /// Calculated as: count_nonzero_activations / total_samples
-    /// where "fires" means |activation| > small threshold (avoiding floating point issues).
-    ///
-    /// This helps identify neurons with extreme firing patterns:
-    /// - activation_frequency < 0.1: Rarely fires, limited influence on most samples
-    /// - activation_frequency > 0.9: Always fires, behaves like a constant (no discriminative power)
-    /// - 0.1 <= activation_frequency <= 0.9: "Sweet spot" with good discriminative power
-    pub activation_frequency: f32,
-}
-
-/// A neuron with activation-weighted impact below removal savings threshold - candidate for removal.
-/// Removing such neurons improves score because complexity reduction outweighs contribution.
-#[derive(Debug)]
-pub struct RemovalCandidate {
-    pub neuron_uuid: String,
-    pub total_error: f32,
-    /// Structural impact based on weight paths to output
-    pub impact: f32,
-    /// Mean absolute activation value from recorded samples
-    pub mean_activation: f32,
-    /// Activation-weighted impact = structural_impact × mean_activation
-    /// This reflects the actual contribution the neuron makes during inference
-    pub activation_weighted_impact: f32,
-    /// Number of synapses pointing TO this neuron
-    pub incoming_synapses: usize,
-    /// Number of synapses pointing FROM this neuron
-    pub outgoing_synapses: usize,
-    /// The complexity savings from removing this neuron (based on NEAT-AI Score.ts formula)
-    pub removal_savings: f32,
-    /// Expected creature-level error reduction from removing this neuron.
-    /// This is based on activation_weighted_impact, NOT the neuron's error.
-    /// Issue #117: Previously, total_error was incorrectly used as expected error reduction,
-    /// leading to predictions like 27% when actual reduction was ~0%.
-    pub expected_error_reduction: f32,
-    pub reason: String,
-}
-
-#[derive(Debug)]
-pub struct RankFocusStats {
-    pub neurons: Vec<RankedNeuron>,
-    /// Neurons with impact below costOfGrowth - candidates for removal
-    pub removal_candidates: Vec<RemovalCandidate>,
-    /// Issue #306: Coordinated structural candidates for removing constant-value neurons.
-    /// When a hidden neuron has near-zero activation variance (constant output), it can be
-    /// removed and its effect folded into bias adjustments for downstream neurons.
-    /// Each candidate contains:
-    /// - A RemoveNeuron operation for the constant neuron
-    /// - SetBias operations for all downstream neurons with adjusted biases
-    pub constant_neuron_removals: Vec<CoordinatedStructuralCandidateJson>,
-    pub max_output_error: f32,
-    pub processed_neurons: usize,
-    pub total_neurons: usize,
-    pub duration_ms: u128,
-}
-
-fn is_selectable_type(neuron_type: &str) -> bool {
-    neuron_type != "input" && neuron_type != "constant"
-}
-
-/// Calculate the complexity savings from removing a neuron.
-///
-/// Based on NEAT-AI's Score.ts formula:
-/// ```typescript
-/// const complexityPenalty = hiddenNeuronCount * growthCost +
-///     creature.synapses.length * growthCost / 10 + penalty * growthCost / 100;
-/// ```
-///
-/// So removing a neuron with N incoming and M outgoing synapses saves:
-/// - `growth_cost` for the neuron itself
-/// - `(N + M) × growth_cost / 10` for the synapses
-///
-/// Total: `growth_cost × (1 + (N + M) / 10)`
-///
-/// # Arguments
-/// * `incoming_synapses` - Number of synapses pointing TO this neuron
-/// * `outgoing_synapses` - Number of synapses pointing FROM this neuron
-/// * `growth_cost` - The cost per hidden neuron (typically 1e-7)
-///
-/// # Returns
-/// The total complexity savings from removing this neuron and its synapses
-pub fn calculate_removal_savings(
-    incoming_synapses: usize,
-    outgoing_synapses: usize,
-    growth_cost: f32,
-) -> f32 {
-    let total_synapses = incoming_synapses + outgoing_synapses;
-    growth_cost * (1.0 + total_synapses as f32 / 10.0)
-}
-
-/// Pre-computed synapse counts for efficient O(1) lookup.
-///
-/// Issue #208: Pre-compute synapse counts to eliminate O(n×m) complexity.
-///
-/// Previously, `count_synapses_for_neuron` performed a linear O(m) scan through ALL synapses
-/// twice (incoming + outgoing) for each neuron being ranked. With n neurons and m synapses,
-/// this was O(n × m) complexity.
-///
-/// This struct pre-builds two HashMaps during initialisation in O(m) time, then provides
-/// O(1) lookup for any neuron's synapse counts. Total complexity is O(n + m).
-///
-/// # Example performance improvement
-///
-/// For a creature with 500 neurons and 10,000 synapses:
-/// - Previous: 500 × 10,000 × 2 = **10 million** iterations
-/// - With SynapseCounts: 10,000 + 500 = **10,500** iterations
-/// - **~1000x improvement**
-#[derive(Debug)]
-pub struct SynapseCounts {
-    /// Map from neuron UUID to count of synapses pointing TO that neuron
-    incoming: HashMap<String, usize>,
-    /// Map from neuron UUID to count of synapses pointing FROM that neuron
-    outgoing: HashMap<String, usize>,
-}
-
-impl SynapseCounts {
-    /// Create a new SynapseCounts by scanning all synapses once.
-    ///
-    /// Time complexity: O(m) where m is the number of synapses.
-    /// Space complexity: O(n) where n is the number of unique neurons with synapses.
-    pub fn new(creature: &CreatureJson) -> Self {
-        let mut incoming: HashMap<String, usize> = HashMap::new();
-        let mut outgoing: HashMap<String, usize> = HashMap::new();
-
-        for synapse in &creature.synapses {
-            *incoming.entry(synapse.to_uuid.clone()).or_default() += 1;
-            *outgoing.entry(synapse.from_uuid.clone()).or_default() += 1;
-        }
-
-        Self { incoming, outgoing }
-    }
-
-    /// Get the synapse counts for a neuron in O(1) time.
-    ///
-    /// # Arguments
-    /// * `neuron_uuid` - The UUID of the neuron to look up
-    ///
-    /// # Returns
-    /// A tuple of (incoming_count, outgoing_count). Returns (0, 0) if the neuron
-    /// has no synapses or doesn't exist in the creature.
-    pub fn get(&self, neuron_uuid: &str) -> (usize, usize) {
-        (
-            self.incoming.get(neuron_uuid).copied().unwrap_or(0),
-            self.outgoing.get(neuron_uuid).copied().unwrap_or(0),
-        )
-    }
-}
-
-fn average_absolute_error_from_records(records: &[DiscoverRecord]) -> f32 {
-    let mut sum = 0.0f32;
-    let mut count: u32 = 0;
-
-    for record in records {
-        for err in &record.errors {
-            if err.is_finite() {
-                sum += err.abs();
-                count += 1;
-            }
-        }
-    }
-
-    if count == 0 { 0.0 } else { sum / count as f32 }
-}
-
-/// Compute mean absolute activation from discovery records.
-/// Sum of |activation| divided by number of finite records.
-///
-/// Non-finite values (NaN, Infinity) are filtered out to prevent
-/// corruption of activation_weighted_impact calculations and sorting.
-fn mean_absolute_activation_from_records(records: &[DiscoverRecord]) -> f32 {
-    if records.is_empty() {
-        return 0.0;
-    }
-
-    let mut sum = 0.0f32;
-    let mut count: u32 = 0;
-
-    for record in records {
-        if record.activation.is_finite() {
-            sum += record.activation.abs();
-            count += 1;
-        }
-    }
-
-    if count == 0 { 0.0 } else { sum / count as f32 }
-}
-
-/// Issue #204: Compute activation frequency from discovery records.
-///
-/// Activation frequency = count_nonzero_activations / total_samples
-/// where "fires" means |activation| > threshold to avoid floating-point issues.
-///
-/// # Arguments
-/// * `records` - Discovery records for a single neuron
-///
-/// # Returns
-/// A value in [0.0, 1.0] representing the proportion of samples where the neuron fires.
-/// Returns 0.0 if there are no valid (finite) records.
-///
-/// # Rationale
-/// - Neurons that rarely fire (frequency < 0.1) have limited influence on most samples
-/// - Neurons that always fire (frequency > 0.9) behave like constants with no discriminative power
-/// - Neurons with moderate frequency (0.1 to 0.9) are in the "sweet spot" for discovery
-const ACTIVATION_FIRING_THRESHOLD: f32 = 1e-6;
-
-fn activation_frequency_from_records(records: &[DiscoverRecord]) -> f32 {
-    if records.is_empty() {
-        return 0.0;
-    }
-
-    let mut firing_count: u32 = 0;
-    let mut total_count: u32 = 0;
-
-    for record in records {
-        if record.activation.is_finite() {
-            total_count += 1;
-            // A neuron "fires" when its absolute activation exceeds the threshold
-            if record.activation.abs() > ACTIVATION_FIRING_THRESHOLD {
-                firing_count += 1;
-            }
-        }
-    }
-
-    if total_count == 0 {
-        0.0
-    } else {
-        firing_count as f32 / total_count as f32
-    }
-}
-
-/// Issue #204: Compute frequency factor for focus neuron ranking.
-///
-/// Neurons with extreme activation frequencies (rarely or always firing) are
-/// less useful for discovery analysis:
-/// - Rarely-firing neurons (< 10%) have limited influence on most samples
-/// - Always-firing neurons (> 90%) behave like constants with no discriminative power
-///
-/// # Arguments
-/// * `activation_frequency` - The proportion of samples where the neuron fires [0.0, 1.0]
-///
-/// # Returns
-/// * 0.8 if activation_frequency < 0.1 (rarely fires) - 20% penalty
-/// * 0.8 if activation_frequency > 0.9 (always fires) - 20% penalty
-/// * 1.0 otherwise (moderate frequency) - no penalty
-const FREQUENCY_LOW_THRESHOLD: f32 = 0.1;
-const FREQUENCY_HIGH_THRESHOLD: f32 = 0.9;
-const FREQUENCY_PENALTY_FACTOR: f32 = 0.8;
-
-fn compute_frequency_factor(activation_frequency: f32) -> f32 {
-    if (FREQUENCY_LOW_THRESHOLD..=FREQUENCY_HIGH_THRESHOLD).contains(&activation_frequency) {
-        1.0 // No penalty for moderate frequency (10-90%)
-    } else {
-        FREQUENCY_PENALTY_FACTOR // 0.8x penalty for extreme frequencies
-    }
-}
-
-/// Compute activation variance and mean from discovery records.
-///
-/// Issue #306: Used to detect constant-value neurons for removal with bias adjustments.
-/// A neuron with near-zero variance has constant activation and can be removed,
-/// with its effect folded into bias adjustments for downstream neurons.
-///
-/// # Returns
-/// A tuple of (mean_activation, variance) where:
-/// - `mean_activation` is the arithmetic mean (NOT absolute value)
-/// - `variance` is the statistical variance of activations
-///
-/// Returns (0.0, 0.0) if there are insufficient records.
-fn activation_mean_and_variance_from_records(records: &[DiscoverRecord]) -> (f32, f32) {
-    if records.len() < 2 {
-        return (0.0, 0.0);
-    }
-
-    let mut sum = 0.0f64;
-    let mut sum_sq = 0.0f64;
-    let mut count = 0u32;
-
-    for record in records {
-        if record.activation.is_finite() {
-            let a = record.activation as f64;
-            sum += a;
-            sum_sq += a * a;
-            count += 1;
-        }
-    }
-
-    if count < 2 {
-        return (0.0, 0.0);
-    }
-
-    let n = count as f64;
-    let mean = sum / n;
-    let variance = (sum_sq / n) - (mean * mean);
-
-    (mean as f32, variance.max(0.0) as f32)
-}
-
-/// Threshold for considering a neuron as "constant" (near-zero variance).
-/// Issue #217/306: Neurons with variance below this threshold are treated as constant
-/// and can be removed with bias adjustments for downstream neurons.
-///
-/// Value 1e-10 is from Issue #217's proposal for DEAD_VARIANCE_THRESHOLD.
-const CONSTANT_VARIANCE_THRESHOLD: f32 = 1e-10;
-
-// =============================================================================
-// Gradient Flow Analysis (Issue #206)
-// =============================================================================
-
-/// Statistics about gradient flow through a neuron.
-///
-/// These metrics help identify neurons with high learning potential by analysing:
-/// 1. How much error signal can flow through the neuron (gradient magnitude)
-/// 2. Whether the neuron is stuck in a saturated region (saturation ratio)
-/// 3. Whether the neuron is effectively dead (dead ratio for ReLU-family)
-///
-/// # Usage in Focus Ranking
-///
-/// Neurons with high gradient magnitude and low saturation/dead ratios have higher
-/// learning potential and should be prioritised for discovery analysis.
-///
-/// # Example
-///
-/// ```ignore
-/// let stats = GradientFlowStats {
-///     avg_gradient_magnitude: 0.75,  // Good gradient flow
-///     saturation_ratio: 0.1,          // 10% saturated samples
-///     dead_ratio: 0.0,                // No dead samples (not ReLU)
-/// };
-/// ```
-#[derive(Debug, Clone, Copy)]
-pub struct GradientFlowStats {
-    /// Average gradient magnitude across all samples.
-    ///
-    /// For a neuron with activation function f(x), this is the average of |f'(x)|
-    /// weighted by the error magnitude. Higher values indicate the neuron can
-    /// effectively propagate error signals for learning.
-    pub avg_gradient_magnitude: f32,
-
-    /// Ratio of samples where the neuron is in a saturated region (0.0 to 1.0).
-    ///
-    /// Saturation thresholds by activation:
-    /// - TANH: |value| > 3 (gradient < 0.01)
-    /// - LOGISTIC: |value| > 5 (output near 0 or 1)
-    /// - HARD_TANH: |value| > 1 (clamped to ±1)
-    ///
-    /// High saturation ratio means the neuron cannot effectively learn because
-    /// the gradient is near zero in the saturated region.
-    pub saturation_ratio: f32,
-
-    /// Ratio of samples where the neuron has zero gradient (0.0 to 1.0).
-    ///
-    /// Primarily relevant for ReLU-family activations:
-    /// - ReLU: value < 0 (completely dead - zero output and gradient)
-    /// - RELU6: value < 0 or value > 6 (dead at both ends)
-    /// - SELU: value < 0 has reduced but non-zero gradient
-    ///
-    /// High dead ratio means the neuron cannot learn from most samples.
-    /// For non-ReLU activations, this will typically be 0.0.
-    pub dead_ratio: f32,
-}
-
-impl Default for GradientFlowStats {
-    fn default() -> Self {
-        Self {
-            avg_gradient_magnitude: 1.0, // Assume full gradient for unknown activations
-            saturation_ratio: 0.0,       // Assume not saturated
-            dead_ratio: 0.0,             // Assume not dead
-        }
-    }
-}
-
-/// Saturation thresholds for various activation functions.
-/// These thresholds define when the gradient becomes effectively zero.
-mod saturation_thresholds {
-    /// TANH: |x| > 3 means tanh'(x) < 0.01 (effectively saturated)
-    pub const TANH_THRESHOLD: f32 = 3.0;
-
-    /// LOGISTIC: |x| > 5 means sigmoid'(x) < 0.007 (effectively saturated)
-    pub const LOGISTIC_THRESHOLD: f32 = 5.0;
-
-    /// HARD_TANH: |x| >= 1 means output is clamped (gradient = 0)
-    pub const HARD_TANH_THRESHOLD: f32 = 1.0;
-
-    /// ARCTAN: |x| > 10 means arctan'(x) < 0.01
-    pub const ARCTAN_THRESHOLD: f32 = 10.0;
-
-    /// SOFTSIGN: |x| > 10 means softsign'(x) < 0.008
-    pub const SOFTSIGN_THRESHOLD: f32 = 10.0;
-
-    /// Gradient threshold below which we consider the sample "saturated"
-    pub const GRADIENT_SATURATION_THRESHOLD: f32 = 0.01;
-}
-
-/// Compute the gradient (derivative) of an activation function at a given input value.
-///
-/// Returns (gradient, is_dead) where:
-/// - gradient: The derivative f'(value)
-/// - is_dead: True if this is a "dead" neuron sample (zero gradient AND zero output)
-///
-/// # Arguments
-/// * `squash` - Uppercase squash function name
-/// * `value` - Pre-activation input value
-fn compute_activation_gradient(squash: &str, value: f32) -> (f32, bool) {
-    match squash {
-        // ReLU family - can have dead neurons
-        "RELU" => {
-            if value > 0.0 {
-                (1.0, false)
-            } else {
-                (0.0, true) // Dead: zero gradient AND zero output
-            }
-        }
-        "RELU6" => {
-            if value <= 0.0 {
-                (0.0, true) // Dead at negative end
-            } else if value >= 6.0 {
-                (0.0, false) // Saturated but not "dead" (output is 6)
-            } else {
-                (1.0, false)
-            }
-        }
-        "LEAKYRELU" => {
-            // LeakyReLU never truly dead - always has 0.01 gradient for negative inputs
-            if value >= 0.0 {
-                (1.0, false)
-            } else {
-                (0.01, false)
-            }
-        }
-        "ELU" => {
-            if value >= 0.0 {
-                (1.0, false)
-            } else {
-                // ELU'(x) = exp(x) for x < 0
-                (value.exp(), false)
-            }
-        }
-        "SELU" => {
-            const ALPHA: f32 = 1.673_263_2;
-            const LAMBDA: f32 = 1.050_701;
-            if value >= 0.0 {
-                (LAMBDA, false)
-            } else {
-                (LAMBDA * ALPHA * value.exp(), false)
-            }
-        }
-
-        // Sigmoid-family - can saturate
-        "TANH" => {
-            let tanh_val = value.tanh();
-            let gradient = 1.0 - tanh_val * tanh_val;
-            (gradient, false)
-        }
-        "LOGISTIC" => {
-            let sigmoid = if value >= 0.0 {
-                1.0 / (1.0 + (-value).exp())
-            } else {
-                let exp_x = value.exp();
-                exp_x / (1.0 + exp_x)
-            };
-            let gradient = sigmoid * (1.0 - sigmoid);
-            (gradient, false)
-        }
-        "HARD_TANH" | "CLIPPED" => {
-            if value.abs() >= saturation_thresholds::HARD_TANH_THRESHOLD {
-                (0.0, false) // Saturated but not "dead"
-            } else {
-                (1.0, false)
-            }
-        }
-        "BIPOLAR_SIGMOID" => {
-            // 2 * sigmoid(x) - 1, derivative = 2 * sigmoid(x) * (1 - sigmoid(x))
-            let sigmoid = if value >= 0.0 {
-                1.0 / (1.0 + (-value).exp())
-            } else {
-                let exp_x = value.exp();
-                exp_x / (1.0 + exp_x)
-            };
-            (2.0 * sigmoid * (1.0 - sigmoid), false)
-        }
-        "SOFTSIGN" => {
-            // softsign'(x) = 1 / (1 + |x|)^2
-            let denom = 1.0 + value.abs();
-            (1.0 / (denom * denom), false)
-        }
-
-        // Other activations
-        "IDENTITY" => (1.0, false),
-        "ABSOLUTE" | "ABS" => {
-            if value != 0.0 {
-                (1.0, false) // |gradient| = 1 everywhere except at 0
-            } else {
-                (0.0, false) // Technically undefined at 0
-            }
-        }
-        "ARCTAN" => {
-            // arctan'(x) = 1 / (1 + x^2)
-            (1.0 / (1.0 + value * value), false)
-        }
-        "BENT_IDENTITY" => {
-            // bent_identity'(x) = x / (2 * sqrt(x^2 + 1)) + 1
-            ((value / (2.0 * (value * value + 1.0).sqrt())) + 1.0, false)
-        }
-        "CUBE" => {
-            // cube'(x) = 3x^2
-            (3.0 * value * value, false)
-        }
-        "SQUARE" => {
-            // square'(x) = 2x
-            (2.0 * value, false)
-        }
-        "GAUSSIAN" => {
-            // gaussian(x) = exp(-x^2), gaussian'(x) = -2x * exp(-x^2)
-            let safe_x = value.abs().min(100.0);
-            ((-2.0 * value * (-safe_x * safe_x).exp()).abs(), false)
-        }
-        "GELU" => {
-            // GELU gradient approximation
-            let x3 = value * value * value;
-            let tanh_arg = 0.797_884_6_f32 * (value + 0.044_715_f32 * x3);
-            let tanh_val = tanh_arg.tanh();
-            let sech2 = 1.0 - tanh_val * tanh_val;
-            let inner_deriv = 0.797_884_6_f32 * (1.0 + 3.0 * 0.044_715_f32 * value * value);
-            let gradient = 0.5 * (1.0 + tanh_val) + 0.5 * value * sech2 * inner_deriv;
-            (gradient, false)
-        }
-        "ISRU" => {
-            // ISRU(x) = x / sqrt(1 + x^2)
-            // ISRU'(x) = 1 / (1 + x^2)^(3/2)
-            let denom = (1.0 + value * value).powf(1.5);
-            (1.0 / denom, false)
-        }
-        "MISH" => {
-            // Mish(x) = x * tanh(softplus(x))
-            // Derivative is complex, approximate with numerical stability
-            let sp = if value > 20.0 {
-                value
-            } else {
-                (1.0 + value.exp()).ln()
-            };
-            let tanh_sp = sp.tanh();
-            let sech2_sp = 1.0 - tanh_sp * tanh_sp;
-            let sp_deriv = if value > 20.0 {
-                1.0
-            } else {
-                value.exp() / (1.0 + value.exp())
-            };
-            (tanh_sp + value * sech2_sp * sp_deriv, false)
-        }
-        "SOFTPLUS" => {
-            // softplus(x) = ln(1 + exp(x)), softplus'(x) = sigmoid(x)
-            let sigmoid = if value >= 0.0 {
-                1.0 / (1.0 + (-value).exp())
-            } else {
-                let exp_x = value.exp();
-                exp_x / (1.0 + exp_x)
-            };
-            (sigmoid, false)
-        }
-        "SWISH" => {
-            // Swish(x) = x * sigmoid(x)
-            // Swish'(x) = sigmoid(x) + x * sigmoid(x) * (1 - sigmoid(x))
-            let sigmoid = if value >= 0.0 {
-                1.0 / (1.0 + (-value).exp())
-            } else {
-                let exp_x = value.exp();
-                exp_x / (1.0 + exp_x)
-            };
-            (sigmoid + value * sigmoid * (1.0 - sigmoid), false)
-        }
-        "SINE" | "SINUSOID" => (value.cos().abs(), false),
-        "COSINE" => (value.sin().abs(), false),
-        "TAN" => {
-            let cos_val = value.cos();
-            if cos_val.abs() < 1e-6 {
-                (100.0, false) // Near asymptote, very high gradient
-            } else {
-                (1.0 / (cos_val * cos_val), false)
-            }
-        }
-        "EXPONENTIAL" => {
-            if value >= 88.0 {
-                (f32::MAX, false)
-            } else {
-                (value.exp(), false)
-            }
-        }
-        "LOGSIGMOID" => {
-            // logsigmoid(x) = -ln(1 + exp(-x))
-            // logsigmoid'(x) = 1 / (1 + exp(x)) = 1 - sigmoid(x)
-            let sigmoid = if value >= 0.0 {
-                1.0 / (1.0 + (-value).exp())
-            } else {
-                let exp_x = value.exp();
-                exp_x / (1.0 + exp_x)
-            };
-            (1.0 - sigmoid, false)
-        }
-        "SQRT" => {
-            if value > 0.0 {
-                (0.5 / value.sqrt(), false)
-            } else {
-                (0.0, false) // Undefined for negative, zero at boundary
-            }
-        }
-        "STDINVERSE" => {
-            // 1/x, derivative = -1/x^2
-            let eps = 1e-15_f32;
-            let safe_x = if value.abs() < eps { eps } else { value.abs() };
-            (1.0 / (safe_x * safe_x), false)
-        }
-        "COMPLEMENT" | "INVERSE" => {
-            // 1 - x, derivative = -1 (absolute value = 1)
-            (1.0, false)
-        }
-
-        // Threshold functions - gradient is zero almost everywhere
-        // but can still propagate useful information at the threshold
-        "STEP" | "BIPOLAR" => {
-            // Technically gradient = 0 everywhere except at threshold
-            // For gradient flow analysis, treat as having small gradient
-            (0.0, false)
-        }
-
-        // Aggregate functions - cannot compute scalar gradient
-        "MINIMUM" | "MAXIMUM" | "IF" | "MEAN" | "HYPOT" | "HYPOTV2" => {
-            // For aggregate functions, assume gradient passes through
-            // The actual gradient depends on which input "wins"
-            (1.0, false)
-        }
-
-        // Unknown activation - assume identity-like behaviour
-        _ => (1.0, false),
-    }
-}
-
-/// Check if a neuron sample is in a saturated region based on its activation function.
-///
-/// Saturation means the gradient is effectively zero, preventing learning.
-fn is_saturated(squash: &str, value: f32, gradient: f32) -> bool {
-    use saturation_thresholds::*;
-
-    // Use function-specific thresholds for common activations
-    let saturated_by_threshold = match squash {
-        "TANH" => value.abs() > TANH_THRESHOLD,
-        "LOGISTIC" => value.abs() > LOGISTIC_THRESHOLD,
-        "HARD_TANH" | "CLIPPED" => value.abs() >= HARD_TANH_THRESHOLD,
-        "ARCTAN" => value.abs() > ARCTAN_THRESHOLD,
-        "SOFTSIGN" => value.abs() > SOFTSIGN_THRESHOLD,
-        "RELU6" => !(0.0..=6.0).contains(&value),
-        // For other functions, rely on gradient threshold
-        _ => false,
-    };
-
-    // Also check gradient magnitude as a fallback
-    saturated_by_threshold || gradient < GRADIENT_SATURATION_THRESHOLD
-}
-
-/// Compute gradient flow statistics for all neurons in a creature.
-///
-/// This function analyses discovery records to compute gradient flow metrics
-/// for each selectable neuron. The metrics help identify neurons with high
-/// learning potential by measuring:
-///
-/// 1. **Gradient magnitude**: Average |f'(value)| across samples
-/// 2. **Saturation ratio**: Fraction of samples in saturated region
-/// 3. **Dead ratio**: Fraction of samples with zero gradient (ReLU-family)
-///
-/// # Arguments
-/// * `parquet_file` - Path to parquet file containing discovery records
-/// * `creature` - The creature to analyse
-///
-/// # Returns
-/// Map from neuron UUID to GradientFlowStats
-///
-/// # Example
-///
-/// ```ignore
-/// let stats = compute_gradient_flow_stats("records.parquet", &creature)?;
-/// for (uuid, flow_stats) in &stats {
-///     println!("{}: gradient={:.3}, saturated={:.1}%, dead={:.1}%",
-///         uuid,
-///         flow_stats.avg_gradient_magnitude,
-///         flow_stats.saturation_ratio * 100.0,
-///         flow_stats.dead_ratio * 100.0
-///     );
-/// }
-/// ```
-pub fn compute_gradient_flow_stats(
-    parquet_file: &str,
-    creature: &CreatureJson,
-) -> Result<HashMap<String, GradientFlowStats>> {
-    // Build squash map for efficient lookup
-    let squash_map = build_squash_map(creature);
-
-    // Load records grouped by neuron
-    let records = read_all_records_grouped_by_neuron(parquet_file)
-        .context("Failed to read discovery records for gradient flow analysis")?;
-
-    // Compute stats for each neuron
-    let stats: HashMap<String, GradientFlowStats> = creature
-        .neurons
-        .iter()
-        .filter(|n| is_selectable_type(&n.neuron_type))
-        .filter_map(|neuron| {
-            let squash = squash_map.get(&neuron.uuid)?;
-            let neuron_records = records.get(&neuron.uuid)?;
-
-            if neuron_records.is_empty() {
-                return Some((neuron.uuid.clone(), GradientFlowStats::default()));
-            }
-
-            let mut total_gradient = 0.0f64;
-            let mut saturated_count = 0u32;
-            let mut dead_count = 0u32;
-            let mut valid_count = 0u32;
-
-            for record in neuron_records.iter() {
-                // Use pre-activation value if available, otherwise skip
-                let value = match record.value {
-                    Some(v) if v.is_finite() => v,
-                    _ => continue,
-                };
-
-                let (gradient, is_dead) = compute_activation_gradient(squash, value);
-
-                if !gradient.is_finite() {
-                    continue;
-                }
-
-                valid_count += 1;
-                total_gradient += gradient as f64;
-
-                if is_dead {
-                    dead_count += 1;
-                }
-
-                if is_saturated(squash, value, gradient) {
-                    saturated_count += 1;
-                }
-            }
-
-            if valid_count == 0 {
-                return Some((neuron.uuid.clone(), GradientFlowStats::default()));
-            }
-
-            let avg_gradient_magnitude = (total_gradient / valid_count as f64) as f32;
-            let saturation_ratio = saturated_count as f32 / valid_count as f32;
-            let dead_ratio = dead_count as f32 / valid_count as f32;
-
-            Some((
-                neuron.uuid.clone(),
-                GradientFlowStats {
-                    avg_gradient_magnitude,
-                    saturation_ratio,
-                    dead_ratio,
-                },
-            ))
-        })
-        .collect();
-
-    Ok(stats)
-}
-
-/// Compute gradient flow stats for a single neuron from its discovery records.
-///
-/// This is a helper function for use within `rank_focus_neurons` where records
-/// are already loaded and we want to avoid re-loading the parquet file.
-///
-/// # Arguments
-/// * `neuron_uuid` - UUID of the neuron
-/// * `squash_map` - Map from neuron UUID to uppercase squash name
-/// * `records` - Discovery records for this neuron
-///
-/// # Returns
-/// GradientFlowStats for the neuron
-fn compute_gradient_flow_for_neuron(
-    neuron_uuid: &str,
-    squash_map: &HashMap<String, String>,
-    records: &[DiscoverRecord],
-) -> GradientFlowStats {
-    let squash = match squash_map.get(neuron_uuid) {
-        Some(s) => s.as_str(),
-        None => return GradientFlowStats::default(),
-    };
-
-    if records.is_empty() {
-        return GradientFlowStats::default();
-    }
-
-    let mut total_gradient = 0.0f64;
-    let mut saturated_count = 0u32;
-    let mut dead_count = 0u32;
-    let mut valid_count = 0u32;
-
-    for record in records {
-        // Use pre-activation value if available, otherwise skip
-        let value = match record.value {
-            Some(v) if v.is_finite() => v,
-            _ => continue,
-        };
-
-        let (gradient, is_dead) = compute_activation_gradient(squash, value);
-
-        if !gradient.is_finite() {
-            continue;
-        }
-
-        valid_count += 1;
-        total_gradient += gradient as f64;
-
-        if is_dead {
-            dead_count += 1;
-        }
-
-        if is_saturated(squash, value, gradient) {
-            saturated_count += 1;
-        }
-    }
-
-    if valid_count == 0 {
-        return GradientFlowStats::default();
-    }
-
-    let avg_gradient_magnitude = (total_gradient / valid_count as f64) as f32;
-    let saturation_ratio = saturated_count as f32 / valid_count as f32;
-    let dead_ratio = dead_count as f32 / valid_count as f32;
-
-    GradientFlowStats {
-        avg_gradient_magnitude,
-        saturation_ratio,
-        dead_ratio,
-    }
-}
-
-/// Compute a ranking factor from gradient flow statistics.
-///
-/// This factor adjusts the base ranking score (error × impact) to prioritise
-/// neurons with high learning potential:
-///
-/// ```text
-/// factor = (1 - saturation_ratio) × (1 - dead_ratio) × gradient_boost
-/// ```
-///
-/// where `gradient_boost = 0.5 + 0.5 × clamp(avg_gradient_magnitude, 0, 1)`
-/// ensures the factor is always positive and rewards higher gradient magnitudes.
-///
-/// # Returns
-/// A factor in range (0.0, 1.0] that multiplies the base ranking score.
-fn compute_gradient_flow_factor(stats: &GradientFlowStats) -> f32 {
-    // Factor for saturation: 1.0 when no saturation, 0.0 when fully saturated
-    let saturation_factor = (1.0 - stats.saturation_ratio).max(0.0);
-
-    // Factor for dead neurons: 1.0 when no dead samples, 0.0 when fully dead
-    let dead_factor = (1.0 - stats.dead_ratio).max(0.0);
-
-    // Factor for gradient magnitude: boost neurons with higher gradient
-    // Clamp to [0, 1] range for consistent scaling, then apply 0.5 + 0.5*x
-    // This ensures factor is always >= 0.5 (never completely kills the score)
-    let clamped_gradient = stats.avg_gradient_magnitude.clamp(0.0, 1.0);
-    let gradient_boost = 0.5 + 0.5 * clamped_gradient;
-
-    // Combine factors multiplicatively
-    // Minimum factor is ~0.01 to ensure neurons aren't completely hidden
-    (saturation_factor * dead_factor * gradient_boost).max(0.01)
-}
-
-fn build_adjacency(creature: &CreatureJson) -> HashMap<String, Vec<(String, f32)>> {
-    let mut adjacency: HashMap<String, Vec<(String, f32)>> = HashMap::new();
-    for synapse in &creature.synapses {
-        adjacency
-            .entry(synapse.from_uuid.clone())
-            .or_default()
-            .push((synapse.to_uuid.clone(), synapse.weight));
-    }
-    adjacency
-}
-
-/// Build a map from neuron UUID to squash function name.
-fn build_squash_map(creature: &CreatureJson) -> HashMap<String, String> {
-    creature
-        .neurons
-        .iter()
-        .map(|n| (n.uuid.clone(), n.squash.to_uppercase()))
-        .collect()
-}
-
-/// Categorise squash functions for impact calculation.
-/// See docs/IMPACT_CALCULATION.md for detailed explanation.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SquashCategory {
-    /// Linear or approximately linear (IDENTITY, TANH, etc.)
-    /// Impact = normalised weight fraction
-    Linear,
-    /// Threshold functions (STEP, BIPOLAR)
-    /// Any input could flip the output - don't normalise
-    Threshold,
-    /// Selection functions (MINIMUM, MAXIMUM)
-    /// Only one synapse "wins" - conservative: don't normalise
-    Selection,
-}
-
-impl SquashCategory {
-    /// Categorise a squash function name.
-    ///
-    /// # Performance (Issue #211)
-    /// Uses `eq_ignore_ascii_case` for zero-allocation case-insensitive comparison.
-    fn from_squash(squash: &str) -> Self {
-        if squash.eq_ignore_ascii_case("STEP") || squash.eq_ignore_ascii_case("BIPOLAR") {
-            Self::Threshold
-        } else if squash.eq_ignore_ascii_case("MINIMUM")
-            || squash.eq_ignore_ascii_case("MAXIMUM")
-            || squash.eq_ignore_ascii_case("IF")
-        {
-            Self::Selection
-        } else {
-            Self::Linear
-        }
-    }
-}
-
-/// Compute selection statistics for MINIMUM, MAXIMUM, and IF neurons using activation records.
-///
-/// For each selection-based neuron, this function analyses the recorded activations to determine
-/// which synapse "wins" (provides the min/max value) for each observation. The result is a map
-/// from (from_uuid, to_uuid) to the probability (0.0 to 1.0) that synapse wins.
-///
-/// For IF neurons with synapse types:
-/// - "condition" synapses: Always contribute, so probability = 1.0
-/// - "positive" synapses: Probability = fraction of observations where condition sum > 0
-/// - "negative" synapses: Probability = fraction of observations where condition sum <= 0
-///
-/// # Arguments
-/// * `creature` - The creature containing neurons and synapses
-/// * `grouped_records` - Activation records grouped by neuron UUID
-///
-/// # Returns
-/// Map from (from_uuid, to_uuid) to win probability for selection-based synapses
-pub fn compute_selection_stats(
-    creature: &CreatureJson,
-    grouped_records: &dyn RecordProvider,
-) -> Result<SelectionStats> {
-    let squash_map = build_squash_map(creature);
-
-    // Find all selection-based neurons (MINIMUM, MAXIMUM, IF)
-    let selection_neurons: Vec<&NeuronJson> = creature
-        .neurons
-        .iter()
-        .filter(|n| {
-            let squash = squash_map.get(&n.uuid).map(|s| s.as_str()).unwrap_or("");
-            matches!(squash, "MINIMUM" | "MAXIMUM" | "IF")
-        })
-        .collect();
-
-    // Process each selection neuron in parallel and collect local stats
-    let partial_stats: Vec<SelectionStats> = selection_neurons
-        .par_iter()
-        .map(|target_neuron| -> Result<Option<SelectionStats>> {
-            let squash = squash_map
-                .get(&target_neuron.uuid)
-                .map(|s| s.as_str())
-                .unwrap_or("");
-
-            // Get incoming synapses to this neuron
-            let incoming_synapses: Vec<&SynapseJson> = creature
-                .synapses
-                .iter()
-                .filter(|s| s.to_uuid == target_neuron.uuid)
-                .collect();
-
-            if incoming_synapses.is_empty() {
-                return Ok(None);
-            }
-
-            let mut local_stats = SelectionStats::new();
-            match squash {
-                "MINIMUM" => {
-                    compute_min_stats(&incoming_synapses, grouped_records, &mut local_stats)?
-                }
-                "MAXIMUM" => {
-                    compute_max_stats(&incoming_synapses, grouped_records, &mut local_stats)?
-                }
-                "IF" => compute_if_stats(&incoming_synapses, grouped_records, &mut local_stats)?,
-                _ => {}
-            }
-            Ok(Some(local_stats))
-        })
-        .collect::<Result<Vec<_>>>()?
-        .into_iter()
-        .flatten()
-        .collect();
-
-    // Merge all partial stats into final result
-    let mut stats = SelectionStats::new();
-    for partial in partial_stats {
-        stats.extend(partial);
-    }
-    Ok(stats)
-}
-
-/// Compute selection statistics for a MINIMUM neuron.
-/// Counts how often each synapse provides the minimum weighted activation.
-fn compute_min_stats(
-    synapses: &[&SynapseJson],
-    grouped_records: &dyn RecordProvider,
-    stats: &mut SelectionStats,
-) -> Result<()> {
-    if synapses.is_empty() {
-        return Ok(());
-    }
-
-    // Build a map of obs_index -> Vec<(synapse_key, weighted_activation)>
-    let mut obs_contributions: ObservationContributions = HashMap::new();
-
-    for synapse in synapses {
-        match grouped_records.get(&synapse.from_uuid)? {
-            None => continue,
-            Some(records) => {
-                for record in records.iter() {
-                    if record.activation.is_finite() {
-                        let weighted = synapse.weight * record.activation;
-                        let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-                        obs_contributions
-                            .entry(record.obs_index)
-                            .or_default()
-                            .push((key, weighted));
-                    }
-                }
-            }
-        }
-    }
-
-    // Count wins for each synapse
-    let mut win_counts: HashMap<(String, String), u32> = HashMap::new();
-    let mut total_obs = 0u32;
-
-    for contributions in obs_contributions.values() {
-        if contributions.is_empty() {
-            continue;
-        }
-        total_obs += 1;
-
-        // Find the minimum weighted activation
-        let min_val = contributions
-            .iter()
-            .map(|(_, v)| *v)
-            .fold(f32::INFINITY, f32::min);
-
-        // Count all synapses that achieved the minimum (handles ties)
-        let winners: Vec<_> = contributions
-            .iter()
-            .filter(|(_, v)| (*v - min_val).abs() < 1e-10)
-            .collect();
-
-        for (key, _) in winners {
-            *win_counts.entry(key.clone()).or_insert(0) += 1;
-        }
-    }
-
-    // Convert counts to probabilities
-    if total_obs > 0 {
-        for synapse in synapses {
-            let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-            let wins = win_counts.get(&key).copied().unwrap_or(0);
-            let probability = wins as f32 / total_obs as f32;
-            stats.insert(key, probability);
-        }
-    }
-    Ok(())
-}
-
-/// Compute selection statistics for a MAXIMUM neuron.
-/// Counts how often each synapse provides the maximum weighted activation.
-fn compute_max_stats(
-    synapses: &[&SynapseJson],
-    grouped_records: &dyn RecordProvider,
-    stats: &mut SelectionStats,
-) -> Result<()> {
-    if synapses.is_empty() {
-        return Ok(());
-    }
-
-    // Build a map of obs_index -> Vec<(synapse_key, weighted_activation)>
-    let mut obs_contributions: ObservationContributions = HashMap::new();
-
-    for synapse in synapses {
-        match grouped_records.get(&synapse.from_uuid)? {
-            None => continue,
-            Some(records) => {
-                for record in records.iter() {
-                    if record.activation.is_finite() {
-                        let weighted = synapse.weight * record.activation;
-                        let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-                        obs_contributions
-                            .entry(record.obs_index)
-                            .or_default()
-                            .push((key, weighted));
-                    }
-                }
-            }
-        }
-    }
-
-    // Count wins for each synapse
-    let mut win_counts: HashMap<(String, String), u32> = HashMap::new();
-    let mut total_obs = 0u32;
-
-    for contributions in obs_contributions.values() {
-        if contributions.is_empty() {
-            continue;
-        }
-        total_obs += 1;
-
-        // Find the maximum weighted activation
-        let max_val = contributions
-            .iter()
-            .map(|(_, v)| *v)
-            .fold(f32::NEG_INFINITY, f32::max);
-
-        // Count all synapses that achieved the maximum (handles ties)
-        let winners: Vec<_> = contributions
-            .iter()
-            .filter(|(_, v)| (*v - max_val).abs() < 1e-10)
-            .collect();
-
-        for (key, _) in winners {
-            *win_counts.entry(key.clone()).or_insert(0) += 1;
-        }
-    }
-
-    // Convert counts to probabilities
-    if total_obs > 0 {
-        for synapse in synapses {
-            let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-            let wins = win_counts.get(&key).copied().unwrap_or(0);
-            let probability = wins as f32 / total_obs as f32;
-            stats.insert(key, probability);
-        }
-    }
-    Ok(())
-}
-
-/// Compute selection statistics for an IF neuron.
-///
-/// IF neurons have three synapse types:
-/// - "condition": Always evaluated to determine which branch to take
-/// - "positive": Used when sum of condition synapses > 0
-/// - "negative": Used when sum of condition synapses <= 0
-///
-/// Impact distribution:
-/// - Condition synapses: probability = 1.0 (always active)
-/// - Positive synapses: probability = fraction of observations where condition > 0
-/// - Negative synapses: probability = fraction of observations where condition <= 0
-fn compute_if_stats(
-    synapses: &[&SynapseJson],
-    grouped_records: &dyn RecordProvider,
-    stats: &mut SelectionStats,
-) -> Result<()> {
-    // Separate synapses by type
-    let condition_synapses: Vec<_> = synapses
-        .iter()
-        .filter(|s| s.synapse_type.as_deref() == Some("condition"))
-        .collect();
-    let positive_synapses: Vec<_> = synapses
-        .iter()
-        .filter(|s| s.synapse_type.as_deref() == Some("positive"))
-        .collect();
-    let negative_synapses: Vec<_> = synapses
-        .iter()
-        .filter(|s| s.synapse_type.as_deref() == Some("negative"))
-        .collect();
-
-    // If no synapse types are set, fall back to equal probability
-    if condition_synapses.is_empty() && positive_synapses.is_empty() && negative_synapses.is_empty()
-    {
-        // No type information - use equal probability fallback
-        let n = synapses.len() as f32;
-        for synapse in synapses {
-            let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-            stats.insert(key, 1.0 / n);
-        }
-        return Ok(());
-    }
-
-    // Condition synapses are always active
-    for synapse in &condition_synapses {
-        let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-        stats.insert(key, 1.0);
-    }
-
-    // Compute condition sum for each observation to determine positive/negative branch usage
-    let mut obs_condition_sums: HashMap<u32, f32> = HashMap::new();
-
-    for synapse in &condition_synapses {
-        match grouped_records.get(&synapse.from_uuid)? {
-            None => continue,
-            Some(records) => {
-                for record in records.iter() {
-                    if record.activation.is_finite() {
-                        let contribution = synapse.weight * record.activation;
-                        *obs_condition_sums.entry(record.obs_index).or_insert(0.0) += contribution;
-                    }
-                }
-            }
-        }
-    }
-
-    // Count positive vs negative branch usage
-    let total_obs = obs_condition_sums.len() as f32;
-    if total_obs == 0.0 {
-        // No observations - use equal probability for positive/negative
-        let pos_count = positive_synapses.len().max(1) as f32;
-        let neg_count = negative_synapses.len().max(1) as f32;
-
-        for synapse in &positive_synapses {
-            let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-            stats.insert(key, 0.5 / pos_count);
-        }
-        for synapse in &negative_synapses {
-            let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-            stats.insert(key, 0.5 / neg_count);
-        }
-        return Ok(());
-    }
-
-    let positive_obs = obs_condition_sums
-        .values()
-        .filter(|&&sum| sum > 0.0)
-        .count() as f32;
-    let negative_obs = total_obs - positive_obs;
-
-    let positive_prob = positive_obs / total_obs;
-    let negative_prob = negative_obs / total_obs;
-
-    // Distribute probability among synapses in each branch
-    // Each synapse in a branch shares that branch's probability equally
-    let pos_count = positive_synapses.len().max(1) as f32;
-    let neg_count = negative_synapses.len().max(1) as f32;
-
-    for synapse in &positive_synapses {
-        let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-        stats.insert(key, positive_prob / pos_count);
-    }
-
-    for synapse in &negative_synapses {
-        let key = (synapse.from_uuid.clone(), synapse.to_uuid.clone());
-        stats.insert(key, negative_prob / neg_count);
-    }
-    Ok(())
-}
-
-// NOTE: build_inbound_weights was removed in v0.1.126 as part of the impact
-// calculation fix. The normalisation it supported was causing massive
-// underestimation of neuron impact (see regression_v0_1_126.rs tests).
-
-/// Public version of compute_impacts for use in add-neuron analysis.
-/// Computes the structural impact of each neuron on outputs (path weight products).
-/// Output neurons have impact = 1.0, hidden neurons have impact in [0, 1] based on
-/// their weighted paths to outputs.
-///
-/// NOTE: This function is squash-aware. For neurons feeding into STEP/BIPOLAR/MINIMUM/MAXIMUM
-/// targets, the impact calculation uses special handling to avoid underestimation.
-/// See docs/IMPACT_CALCULATION.md for detailed explanation.
-pub fn compute_impacts_public(creature: &CreatureJson) -> HashMap<String, f32> {
-    compute_impacts_internal(creature)
-}
-
-/// Context for impact calculation, containing pre-computed lookup tables.
-/// This struct groups related parameters to avoid clippy::too_many_arguments.
-struct ImpactContext {
-    adjacency: HashMap<String, Vec<(String, f32)>>,
-    inbound_count: HashMap<String, usize>,
-    /// Sum of |weight| for all synapses INTO each target neuron.
-    /// Used for normalising Linear squash impact: |w| / total_inbound_weight × child_impact
-    /// This ensures hidden neurons always have impact < 1.0 (Issue #130).
-    total_inbound_weight: HashMap<String, f32>,
-    squash_map: HashMap<String, String>,
-    outputs: HashSet<String>,
-    /// Selection statistics from activation records.
-    /// When available, provides actual win probabilities for MIN/MAX/IF synapses
-    /// instead of the conservative 1/N equal probability fallback.
-    selection_stats: Option<SelectionStats>,
-}
-
-fn compute_impacts_internal(creature: &CreatureJson) -> HashMap<String, f32> {
-    compute_impacts_internal_with_stats(creature, None)
-        .expect("impact computation without records should not fail")
-}
-
-/// Compute impacts with optional activation-based selection statistics.
-///
-/// When `grouped_records` is provided, computes actual selection probabilities for
-/// MIN/MAX/IF neurons based on recorded activations. This gives more accurate
-/// impact estimates than the conservative 1/N equal probability fallback.
-///
-/// # Arguments
-/// * `creature` - The creature to compute impacts for
-/// * `grouped_records` - Optional activation records grouped by neuron UUID
-///
-/// # Returns
-/// Map from neuron UUID to impact score
-fn compute_impacts_internal_with_stats(
-    creature: &CreatureJson,
-    grouped_records: Option<&dyn RecordProvider>,
-) -> Result<HashMap<String, f32>> {
-    let adjacency = build_adjacency(creature);
-    let squash_map = build_squash_map(creature);
-
-    // Build inbound synapse count for selection squashes (MIN/MAX/IF neurons)
-    let inbound_count: HashMap<String, usize> = {
-        let mut map: HashMap<String, usize> = HashMap::new();
-        for synapse in &creature.synapses {
-            *map.entry(synapse.to_uuid.clone()).or_insert(0) += 1;
-        }
-        map
-    };
-
-    // Build total inbound weight for Linear squash normalisation (Issue #130).
-    // Sum of |weight| for all synapses INTO each target neuron.
-    // This ensures hidden neurons always have impact < 1.0:
-    //   contribution = |weight| / total_inbound_weight × child_impact
-    let total_inbound_weight: HashMap<String, f32> = {
-        let mut map: HashMap<String, f32> = HashMap::new();
-        for synapse in &creature.synapses {
-            *map.entry(synapse.to_uuid.clone()).or_insert(0.0) += synapse.weight.abs();
-        }
-        map
-    };
-
-    let outputs: HashSet<String> = creature
-        .neurons
-        .iter()
-        .filter(|n| n.neuron_type == "output")
-        .map(|n| n.uuid.clone())
-        .collect();
-
-    // Compute selection statistics if activation records are available
-    let selection_stats = grouped_records
-        .map(|records| compute_selection_stats(creature, records))
-        .transpose()?;
-
-    let ctx = ImpactContext {
-        adjacency,
-        inbound_count,
-        total_inbound_weight,
-        squash_map,
-        outputs,
-        selection_stats,
-    };
-
-    // Parallel impact computation using thread-local caches
-    // Each thread computes impacts for a subset of neurons, then we merge results.
-    // This trades some redundant computation for better CPU utilization.
-    //
-    // Note: We include ALL neurons (including inputs) to compute their impacts.
-    // Input neurons have outgoing synapses and their impact measures their
-    // contribution to the final output. This is useful for visualisation/debugging.
-    let all_neurons: Vec<&NeuronJson> = creature.neurons.iter().collect();
-
-    // Use a shared cache protected by a mutex for thread-safe updates
-    let shared_cache: Mutex<HashMap<String, f32>> = Mutex::new(HashMap::new());
-
-    all_neurons.par_iter().for_each(|neuron| {
-        // Check if already computed (another thread might have done it)
-        {
-            let cache = shared_cache.lock().unwrap();
-            if cache.contains_key(&neuron.uuid) {
-                return;
-            }
-        }
-
-        // Compute with a local visiting set (cycle detection is per-path)
-        let mut visiting = HashSet::new();
-
-        // We need to compute recursively, but with shared cache access
-        let impact =
-            compute_impact_with_shared_cache(&neuron.uuid, &ctx, &shared_cache, &mut visiting);
-
-        // Store result
-        let mut cache = shared_cache.lock().unwrap();
-        cache.insert(neuron.uuid.clone(), impact);
-    });
-
-    Ok(shared_cache.into_inner().unwrap())
-}
-
-/// Compute impact with a shared cache for parallel execution.
-fn compute_impact_with_shared_cache(
-    uuid: &str,
-    ctx: &ImpactContext,
-    shared_cache: &Mutex<HashMap<String, f32>>,
-    visiting: &mut HashSet<String>,
-) -> f32 {
-    // Check cache first
-    {
-        let cache = shared_cache.lock().unwrap();
-        if let Some(&value) = cache.get(uuid) {
-            return value;
-        }
-    }
-
-    if !visiting.insert(uuid.to_string()) {
-        // Cycle detected; treat as zero contribution
-        return 0.0;
-    }
-
-    let impact = if ctx.outputs.contains(uuid) {
-        1.0
-    } else if let Some(edges) = ctx.adjacency.get(uuid) {
-        // Sum across all outgoing edges
-        let mut total_impact = 0.0;
-        for (to_uuid, weight) in edges {
-            let child_impact =
-                compute_impact_with_shared_cache(to_uuid, ctx, shared_cache, visiting);
-            if child_impact <= 0.0 {
-                continue;
-            }
-
-            let squash = ctx
-                .squash_map
-                .get(to_uuid)
-                .map(|s| s.as_str())
-                .unwrap_or("IDENTITY");
-            let category = SquashCategory::from_squash(squash);
-
-            let contribution = match category {
-                SquashCategory::Linear => {
-                    // Issue #130: Normalise by total inbound weight to ensure hidden neurons
-                    // always have impact < 1.0. This matches the documented formula:
-                    //   contribution = |weight| / total_inbound_weight × child_impact
-                    //
-                    // Without normalisation, a hidden neuron with weight 3.0 to an output
-                    // would get impact = 3.0, which is mathematically incorrect for the
-                    // PURPOSE of prediction discounting (measuring fraction of influence).
-                    //
-                    // Edge case: If all inbound weights are 0.0, total is 0.0, and we'd get
-                    // 0.0 / 0.0 = NaN. Handle this by returning 0.0 (zero weight = zero contribution).
-                    //
-                    // Near-zero protection: The `.max(weight.abs())` ensures total >= weight.abs(),
-                    // so the ratio weight.abs() / total is ALWAYS in [0, 1] and cannot explode.
-                    // Example: weight=1e-10, total_inbound=1e-10 → total=1e-10 → ratio=1.0 ✓
-                    // The only problematic case is weight=0.0 AND total=0.0 → 0/0=NaN, handled below.
-                    let total = ctx
-                        .total_inbound_weight
-                        .get(to_uuid)
-                        .copied()
-                        .unwrap_or(1.0)
-                        .max(weight.abs()); // Bounds ratio to [0,1]: total >= weight.abs() always
-
-                    if total <= 0.0 {
-                        // All weights are zero (including this one) → zero contribution
-                        0.0
-                    } else {
-                        (weight.abs() / total) * child_impact
-                    }
-                }
-                SquashCategory::Threshold => child_impact,
-                SquashCategory::Selection => {
-                    if let Some(ref stats) = ctx.selection_stats {
-                        let key = (uuid.to_string(), to_uuid.clone());
-                        let win_prob = stats.get(&key).copied().unwrap_or_else(|| {
-                            let n = ctx.inbound_count.get(to_uuid).copied().unwrap_or(1).max(1);
-                            1.0 / n as f32
-                        });
-                        win_prob * child_impact
-                    } else {
-                        let n = ctx.inbound_count.get(to_uuid).copied().unwrap_or(1).max(1);
-                        (1.0 / n as f32) * child_impact
-                    }
-                }
-            };
-            total_impact += contribution;
-        }
-        total_impact
-    } else {
-        0.0
-    };
-
-    visiting.remove(uuid);
-
-    // Cache the result
-    {
-        let mut cache = shared_cache.lock().unwrap();
-        cache.insert(uuid.to_string(), impact);
-    }
-
-    impact
-}
-
-/// Public version that computes impacts with activation-based selection statistics.
-///
-/// When activation records are provided, this function computes actual win probabilities
-/// for MIN/MAX/IF neurons instead of using the conservative 1/N equal probability.
-/// This results in more accurate impact estimates.
-///
-/// # Arguments
-/// * `creature` - The creature to compute impacts for
-/// * `grouped_records` - Activation records grouped by neuron UUID
-///
-/// # Returns
-/// Map from neuron UUID to impact score
-pub fn compute_impacts_with_activations(
-    creature: &CreatureJson,
-    grouped_records: &dyn RecordProvider,
-) -> Result<HashMap<String, f32>> {
-    compute_impacts_internal_with_stats(creature, Some(grouped_records))
-}
-
-pub fn rank_focus_neurons(
-    parquet_file: &str,
-    creature: &CreatureJson,
-    max_results: Option<usize>,
-    cost_of_growth: Option<f32>,
-) -> Result<RankFocusStats> {
-    let start = Instant::now();
-    let selectable: Vec<&NeuronJson> = creature
-        .neurons
-        .iter()
-        .filter(|neuron| is_selectable_type(&neuron.neuron_type))
-        .collect();
-    let total_neurons = selectable.len();
-
-    if total_neurons == 0 {
-        return Ok(RankFocusStats {
-            neurons: Vec::new(),
-            removal_candidates: Vec::new(),
-            constant_neuron_removals: Vec::new(),
-            max_output_error: 0.0,
-            processed_neurons: 0,
-            total_neurons: 0,
-            duration_ms: start.elapsed().as_millis(),
-        });
-    }
-
-    // Try to load all records if we have enough memory.
-    // If not enough memory, fall back to lazy loading with a bounded cache.
-    let (records_provider, is_lazy_mode): (Arc<dyn RecordProvider>, bool) =
-        match check_memory_for_parquet(parquet_file) {
-            Ok(()) => {
-                let records = read_all_records_grouped_by_neuron(parquet_file)
-                    .context("Failed to read discovery records from parquet file")?;
-                (Arc::new(EagerRecordProvider::new(records)), false)
-            }
-            Err(memory_error) => {
-                eprintln!(
-                    "[NEAT-AI-Discovery] Insufficient memory for full pre-load in focus ranking. \
-                     Using lazy-loading mode (slower but memory-efficient)."
-                );
-                if verbose_enabled() {
-                    eprintln!("[NEAT-AI-Discovery][verbose] Memory check failed: {memory_error}");
-                }
-                (Arc::new(LazyRecordProvider::new(parquet_file)), true)
-            }
-        };
-
-    if is_lazy_mode && verbose_enabled() {
-        eprintln!(
-            "[NEAT-AI-Discovery][verbose] Lazy record cache initialised (cached: {} neurons)",
-            records_provider.len()
-        );
-    }
-
-    // Verify that all selectable neurons have records (restore old error behaviour)
-    for neuron in &selectable {
-        get_records_or_error(records_provider.as_ref(), &neuron.uuid)
-            .context("Failed to read discovery records for all selectable neurons")?;
-    }
-
-    let output_neurons: Vec<&NeuronJson> = creature
-        .neurons
-        .iter()
-        .filter(|neuron| neuron.neuron_type == "output")
-        .collect();
-
-    let max_output_error = if output_neurons.is_empty() {
-        0.0
-    } else {
-        let errors: Vec<f32> = output_neurons
-            .iter()
-            .map(|neuron| {
-                let records = get_records_or_error(records_provider.as_ref(), &neuron.uuid)?;
-                Ok(average_absolute_error_from_records(&records))
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        errors.into_iter().fold(0.0, f32::max)
-    };
-
-    // Use activation-based impact calculation for more accurate MIN/MAX/IF statistics
-    let impact_map = compute_impacts_with_activations(creature, records_provider.as_ref())?;
-
-    // Issue #208: Pre-compute synapse counts to eliminate O(n×m) complexity.
-    // Previously, count_synapses_for_neuron scanned ALL synapses twice (incoming + outgoing)
-    // for each neuron being ranked. With n neurons and m synapses, this was O(n × m).
-    // By pre-computing counts into HashMaps, we reduce to O(m) for init + O(1) per lookup.
-    let synapse_counts = SynapseCounts::new(creature);
-
-    // Issue #206: Build squash map for gradient flow analysis
-    let squash_map = build_squash_map(creature);
-
-    // Now we can safely unwrap since we've verified all selectable neurons have records
-    // Use parallel iteration for faster processing on multi-core systems
-    let mut neurons: Vec<RankedNeuron> = selectable
-        .par_iter()
-        .map(|neuron| -> Result<RankedNeuron> {
-            let records = get_records_or_error(records_provider.as_ref(), &neuron.uuid)?;
-            let raw_error = average_absolute_error_from_records(&records);
-            let structural_impact = *impact_map.get(&neuron.uuid).unwrap_or(&0.0);
-            let mean_activation = mean_absolute_activation_from_records(&records);
-
-            // Activation-weighted impact reflects the ACTUAL contribution during inference.
-            // A neuron with tiny structural impact but massive activations still contributes
-            // significantly: actual_contribution ≈ weight × activation
-            //
-            // Only neurons with BOTH low structural impact AND low activation should be
-            // removal candidates. If either is high, the neuron is contributing.
-            let activation_weighted_impact = structural_impact * mean_activation;
-
-            let total_error = if max_output_error > 0.0 {
-                raw_error.min(max_output_error)
-            } else {
-                raw_error
-            };
-
-            // Issue #206: Compute gradient flow stats for this neuron
-            let gradient_flow =
-                compute_gradient_flow_for_neuron(&neuron.uuid, &squash_map, &records);
-
-            // Issue #204: Compute activation frequency for focus neuron ranking
-            let activation_frequency = activation_frequency_from_records(&records);
-
-            Ok(RankedNeuron {
-                neuron_uuid: neuron.uuid.clone(),
-                total_error,
-                raw_error,
-                impact: structural_impact,
-                mean_activation,
-                activation_weighted_impact,
-                gradient_flow,
-                activation_frequency,
-            })
-        })
-        .collect::<Result<Vec<_>>>()?;
-
-    // Sort by weighted score (error × impact × gradient_factor × frequency_factor) to prioritise neurons that:
-    // 1. Have high error (potential for improvement)
-    // 2. Have high impact (changes will affect output)
-    // 3. Have good gradient flow (can actually learn - Issue #206)
-    //
-    // Dec 2025: We deliberately soften (but do not remove) the output bias by applying a
-    // sub-linear exponent to impact. This increases exploration of hidden neurons without
-    // letting low-impact neurons dominate purely due to noisy per-neuron errors.
-    //
-    // Jan 2026 (Issue #206): We further adjust ranking by gradient flow factor:
-    // - Neurons stuck in saturation (high saturation_ratio) are de-prioritised
-    // - Dead ReLU neurons (high dead_ratio) are de-prioritised
-    // - Neurons with good gradient flow get higher priority
-    //
-    // Jan 2026 (Issue #204): We also adjust ranking by activation frequency factor:
-    // - Rarely-firing neurons (< 10% activation rate) are de-prioritised (0.8x penalty)
-    // - Always-firing neurons (> 90% activation rate) are de-prioritised (0.8x penalty)
-    // - Moderate-frequency neurons (10-90%) get no penalty
-    const IMPACT_EPSILON: f32 = 0.0001;
-    const IMPACT_GAMMA: f32 = 0.8;
-    neurons.sort_by(|a, b| {
-        // Base weighted score: error × impact^gamma
-        let a_base = a.total_error * (a.impact + IMPACT_EPSILON).powf(IMPACT_GAMMA);
-        let b_base = b.total_error * (b.impact + IMPACT_EPSILON).powf(IMPACT_GAMMA);
-
-        // Issue #206: Apply gradient flow factor
-        // Factor = (1 - saturation_ratio) × (1 - dead_ratio) × (0.5 + avg_gradient_magnitude/2)
-        // This gives higher scores to neurons with:
-        // - Low saturation (more room to learn)
-        // - Low dead ratio (not stuck at zero)
-        // - Higher gradient magnitude (better error propagation)
-        let a_gradient_factor = compute_gradient_flow_factor(&a.gradient_flow);
-        let b_gradient_factor = compute_gradient_flow_factor(&b.gradient_flow);
-
-        // Issue #204: Apply activation frequency factor
-        // Penalises neurons that rarely fire (< 10%) or always fire (> 90%)
-        let a_frequency_factor = compute_frequency_factor(a.activation_frequency);
-        let b_frequency_factor = compute_frequency_factor(b.activation_frequency);
-
-        let a_weighted = a_base * a_gradient_factor * a_frequency_factor;
-        let b_weighted = b_base * b_gradient_factor * b_frequency_factor;
-
-        b_weighted
-            .total_cmp(&a_weighted)
-            .then_with(|| b.impact.total_cmp(&a.impact))
-            .then_with(|| a.neuron_uuid.cmp(&b.neuron_uuid))
-    });
-
-    // Identify removal candidates: neurons with activation_weighted_impact < costOfGrowth.
-    //
-    // activation_weighted_impact = structural_impact × mean_activation
-    // where structural_impact = NORMALISED impact through the network (Issue #130 fix)
-    //
-    // With normalised impact, structural_impact represents the FRACTION of influence
-    // a neuron has on outputs (attribution), always in range [0, 1] per output.
-    // This correctly represents how much removing the neuron would affect results.
-    //
-    // Neurons with impact below costOfGrowth are net negative - removing them
-    // reduces complexity more than it affects error.
-    //
-    // We provide complexity savings info for each neuron based on NEAT-AI's formula:
-    //   savings = growthCost × (1 + (N + M) / 10)
-    // where N = incoming synapses, M = outgoing synapses
-    //
-    // Issue #132: costOfGrowth should be passed from NEAT-AI, not hardcoded.
-    // The correct default is 1e-7 (per hidden neuron) as per NEAT-AI's Score.ts formula.
-    // v0.1.145 incorrectly changed this to 0.01 which caused 418 false removal candidates.
-    const DEFAULT_COST_OF_GROWTH: f32 = 1e-7;
-    let cost_of_growth_threshold = cost_of_growth.unwrap_or(DEFAULT_COST_OF_GROWTH);
-
-    // Issue #235: Return ALL neurons where removal improves the creature's score.
-    //
-    // The only filter is: "will this candidate improve the creature's score?"
-    // A removal improves score when: removal_savings > activation_weighted_impact
-    //
-    // Previously, we filtered on activation_weighted_impact < cost_of_growth_threshold,
-    // but this missed neurons where removal_savings (due to many synapses) exceeds
-    // the neuron's contribution even when activation_weighted_impact is above threshold.
-    //
-    // Use parallel iteration for faster processing on multi-core systems
-    let mut removal_candidates: Vec<RemovalCandidate> = neurons
-        .par_iter()
-        .filter_map(|n| {
-            // Issue #208: Use pre-computed synapse counts for O(1) lookup
-            let (incoming, outgoing) = synapse_counts.get(&n.neuron_uuid);
-            let savings = calculate_removal_savings(incoming, outgoing, cost_of_growth_threshold);
-
-            // Issue #235: Filter on savings > impact (removal improves score)
-            // instead of impact < threshold (may miss valid candidates)
-            if savings <= n.activation_weighted_impact {
-                return None;
-            }
-
-            // Issue #117: expected_error_reduction should be based on activation_weighted_impact,
-            // NOT total_error. The activation_weighted_impact represents the actual contribution
-            // this neuron makes to the output. Removing it changes error by approximately this amount.
-            //
-            // Previously, total_error (neuron's average error) was incorrectly used, leading to
-            // predictions like 27% when actual reduction was ~0% (for low-impact neurons).
-            let expected_error_reduction = n.activation_weighted_impact;
-
-            // Net score improvement = savings - impact
-            let net_improvement = savings - n.activation_weighted_impact;
-
-            Some(RemovalCandidate {
-                neuron_uuid: n.neuron_uuid.clone(),
-                total_error: n.total_error,
-                impact: n.impact,
-                mean_activation: n.mean_activation,
-                activation_weighted_impact: n.activation_weighted_impact,
-                incoming_synapses: incoming,
-                outgoing_synapses: outgoing,
-                removal_savings: savings,
-                expected_error_reduction,
-                reason: format!(
-                    "Removal improves score: saves {:.2e} > impact {:.2e} (net +{:.2e}), {} synapses, costOfGrowth={:.2e}",
-                    savings,
-                    n.activation_weighted_impact,
-                    net_improvement,
-                    incoming + outgoing,
-                    cost_of_growth_threshold,
-                ),
-            })
-        })
-        .collect();
-
-    // Issue #414: High-error exploratory ablation DISABLED
-    //
-    // Previously, neurons with raw_error >= 10× max_output_error were returned as
-    // "exploratory ablation candidates". This discovery type had a 0% success rate
-    // (0 successes from 2 attempts) because the fundamental assumption was flawed:
-    //
-    // **High error ≠ harmful neuron**
-    //
-    // A neuron with high recorded error is often:
-    // 1. Handling the most difficult samples (it's the only path for hard cases)
-    // 2. Receiving bad inputs from upstream (the error is a symptom, not a cause)
-    // 3. Fighting against incorrect biases elsewhere in the network
-    //
-    // Removing such neurons typically makes performance WORSE because:
-    // - The difficult samples lose their only computation path
-    // - The network loses the only neuron attempting to handle a specific pattern
-    //
-    // Error magnitude measures how WRONG the neuron's output is, not how HARMFUL
-    // the neuron is to the network's overall score. This is why predicted
-    // improvements (based on error magnitude) did not match actual outcomes.
-    //
-    // The legitimate removal candidate detection (based on activation_weighted_impact
-    // < costOfGrowth) remains active and has a 17.6% success rate.
-
-    // Issue #235: Sort by net improvement (removal_savings - activation_weighted_impact).
-    // Higher net improvement = better candidate (removing it saves more than its contribution).
-    removal_candidates.sort_by(|a, b| {
-        // Calculate net improvement for each candidate
-        let a_net = a.removal_savings - a.activation_weighted_impact;
-        let b_net = b.removal_savings - b.activation_weighted_impact;
-
-        // Sort by descending net improvement (best candidates first)
-        b_net
-            .total_cmp(&a_net)
-            .then_with(|| {
-                // For ties, prefer lower impact (safer removal)
-                a.activation_weighted_impact
-                    .total_cmp(&b.activation_weighted_impact)
-            })
-            .then_with(|| a.neuron_uuid.cmp(&b.neuron_uuid))
-    });
-
-    if let Some(limit) = max_results
-        && neurons.len() > limit
-    {
-        neurons.truncate(limit);
-    }
-
-    // Issue #306: Detect constant-value neurons and create coordinated structural candidates
-    // that remove the neuron and adjust downstream biases.
-    //
-    // A neuron with near-zero activation variance is "constant" - it always outputs roughly
-    // the same value regardless of input. Removing it is equivalent to adjusting the biases
-    // of downstream neurons by: bias_adjustment = synapse_weight × mean_activation
-    //
-    // This is a win because:
-    // 1. We reduce complexity (one less neuron and its synapses)
-    // 2. We preserve the network's behaviour (downstream biases compensate)
-    // 3. The constant neuron wasn't contributing useful signal anyway
-    let neuron_types: HashMap<&str, &str> = creature
-        .neurons
-        .iter()
-        .map(|n| (n.uuid.as_str(), n.neuron_type.as_str()))
-        .collect();
-
-    let neuron_biases: HashMap<&str, f32> = creature
-        .neurons
-        .iter()
-        .map(|n| (n.uuid.as_str(), n.bias))
-        .collect();
-
-    // Build outgoing synapse map: from_uuid -> [(to_uuid, weight)]
-    let outgoing_synapses: HashMap<&str, Vec<(&str, f32)>> =
-        creature
-            .synapses
-            .iter()
-            .fold(HashMap::new(), |mut map, syn| {
-                map.entry(syn.from_uuid.as_str())
-                    .or_default()
-                    .push((syn.to_uuid.as_str(), syn.weight));
-                map
-            });
-
-    // Find constant hidden neurons and create coordinated removal candidates
-    let constant_neuron_removals: Vec<CoordinatedStructuralCandidateJson> = selectable
-        .par_iter()
-        .filter_map(|neuron| {
-            // Only consider hidden neurons for constant removal
-            let neuron_type = neuron_types.get(neuron.uuid.as_str())?;
-            if *neuron_type != "hidden" {
-                return None;
-            }
-
-            // Get records and compute variance
-            let records = get_records_or_error(records_provider.as_ref(), &neuron.uuid).ok()?;
-            let (mean_activation, variance) = activation_mean_and_variance_from_records(&records);
-
-            // Check if variance is below threshold (constant neuron)
-            if variance > CONSTANT_VARIANCE_THRESHOLD {
-                return None;
-            }
-
-            // Get outgoing synapses for this neuron
-            let outgoing = outgoing_synapses.get(neuron.uuid.as_str())?;
-            if outgoing.is_empty() {
-                return None;
-            }
-
-            // Build coordinated structural candidate:
-            // 1. SetBias operations for all downstream neurons
-            // 2. RemoveNeuron operation
-            let mut operations = Vec::with_capacity(outgoing.len() + 1);
-
-            // Add SetBias operations for all downstream neurons
-            for (to_uuid, weight) in outgoing {
-                let old_bias = neuron_biases.get(to_uuid).copied().unwrap_or(0.0);
-                let bias_adjustment = weight * mean_activation;
-                let new_bias = old_bias + bias_adjustment;
-
-                if new_bias.is_finite() {
-                    operations.push(CoordinatedStructuralOpJson::SetBias {
-                        neuron_uuid: to_uuid.to_string(),
-                        bias: new_bias,
-                    });
-                }
-            }
-
-            // Add RemoveNeuron operation (must be last so bias adjustments happen first)
-            operations.push(CoordinatedStructuralOpJson::RemoveNeuron {
-                neuron_uuid: neuron.uuid.clone(),
-            });
-
-            // Calculate expected improvement: removal savings (complexity reduction)
-            // Issue #208: Use pre-computed synapse counts for O(1) lookup
-            let (incoming_count, outgoing_count) = synapse_counts.get(&neuron.uuid);
-            let removal_savings =
-                calculate_removal_savings(incoming_count, outgoing_count, cost_of_growth_threshold);
-
-            Some(CoordinatedStructuralCandidateJson {
-                operations,
-                expected_creature_score_gain: removal_savings,
-                comment: Some(format!(
-                    "Issue #306: Constant neuron removal with bias adjustments. \
-                     mean_activation={mean_activation:.6}, variance={variance:.2e}, \
-                     {} downstream neurons, savings={removal_savings:.2e}",
-                    outgoing.len()
-                )),
-            })
-        })
-        .collect();
-
-    Ok(RankFocusStats {
-        neurons,
-        removal_candidates,
-        constant_neuron_removals,
-        max_output_error,
-        processed_neurons: total_neurons,
-        total_neurons,
-        duration_ms: start.elapsed().as_millis(),
-    })
-}
-
-/// Rank focus neurons with optional historical discovery success data.
-///
-/// Issue #227: By tracking which neurons have historically led to successful discoveries
-/// (candidates that survived ablation testing), we can prioritise them in future runs,
-/// improving the discovery hit rate.
-///
-/// This function behaves identically to `rank_focus_neurons` when no history is provided.
-/// When history is provided, the ranking score is adjusted to favour neurons with
-/// higher historical success rates using a Bayesian scoring approach:
-///
-/// ```text
-/// combined_score = base_score × history_factor
-/// ```
-///
-/// where:
-/// - `base_score` = error × impact^gamma (same as rank_focus_neurons)
-/// - `history_factor` = bayesian_score from history (0.0 to 1.0)
-/// - For neurons not in history, `history_factor` = 0.5 (neutral prior)
-///
-/// # Arguments
-///
-/// * `parquet_file` - Path to the parquet file containing discovery records
-/// * `creature` - The creature to rank neurons for
-/// * `max_results` - Optional maximum number of neurons to return
-/// * `cost_of_growth` - Optional cost of growth threshold (default: 1e-7)
-/// * `history` - Optional discovery history for historical success data
-///
-/// # Returns
-///
-/// Returns `RankFocusStats` with neurons sorted by combined score (error × impact × history).
-///
-/// # Example
-///
-/// ```ignore
-/// use neat_ai_discovery::discovery_history::DiscoveryHistory;
-/// use neat_ai_discovery::focus::rank_focus_neurons_with_history;
-///
-/// // Create history from previous runs
-/// let mut history = DiscoveryHistory::new();
-/// history.record("hidden-1", true, Some(epoch));  // Success
-/// history.record("hidden-2", false, None);         // Failure
-///
-/// // Rank neurons, prioritising those with higher historical success
-/// let result = rank_focus_neurons_with_history(
-///     "records.parquet",
-///     &creature,
-///     Some(10),      // max_results
-///     None,          // cost_of_growth (use default)
-///     Some(&history),
-/// )?;
-/// ```
-pub fn rank_focus_neurons_with_history(
-    parquet_file: &str,
-    creature: &CreatureJson,
-    max_results: Option<usize>,
-    cost_of_growth: Option<f32>,
-    history: Option<&DiscoveryHistory>,
-) -> Result<RankFocusStats> {
-    let start = Instant::now();
-    let selectable: Vec<&NeuronJson> = creature
-        .neurons
-        .iter()
-        .filter(|neuron| is_selectable_type(&neuron.neuron_type))
-        .collect();
-    let total_neurons = selectable.len();
-
-    if total_neurons == 0 {
-        return Ok(RankFocusStats {
-            neurons: Vec::new(),
-            removal_candidates: Vec::new(),
-            constant_neuron_removals: Vec::new(),
-            max_output_error: 0.0,
-            processed_neurons: 0,
-            total_neurons: 0,
-            duration_ms: start.elapsed().as_millis(),
-        });
-    }
-
-    // Try to load all records if we have enough memory.
-    // If not enough memory, fall back to lazy loading with a bounded cache.
-    let (records_provider, is_lazy_mode): (Arc<dyn RecordProvider>, bool) =
-        match check_memory_for_parquet(parquet_file) {
-            Ok(()) => {
-                let records = read_all_records_grouped_by_neuron(parquet_file)
-                    .context("Failed to read discovery records from parquet file")?;
-                (Arc::new(EagerRecordProvider::new(records)), false)
-            }
-            Err(memory_error) => {
-                eprintln!(
-                    "[NEAT-AI-Discovery] Insufficient memory for full pre-load in focus ranking. \
-                     Using lazy-loading mode (slower but memory-efficient)."
-                );
-                if verbose_enabled() {
-                    eprintln!("[NEAT-AI-Discovery][verbose] Memory check failed: {memory_error}");
-                }
-                (Arc::new(LazyRecordProvider::new(parquet_file)), true)
-            }
-        };
-
-    if is_lazy_mode && verbose_enabled() {
-        eprintln!(
-            "[NEAT-AI-Discovery][verbose] Lazy record cache initialised (cached: {} neurons)",
-            records_provider.len()
-        );
-    }
-
-    // Verify that all selectable neurons have records
-    for neuron in &selectable {
-        get_records_or_error(records_provider.as_ref(), &neuron.uuid)
-            .context("Failed to read discovery records for all selectable neurons")?;
-    }
-
-    let output_neurons: Vec<&NeuronJson> = creature
-        .neurons
-        .iter()
-        .filter(|neuron| neuron.neuron_type == "output")
-        .collect();
-
-    let max_output_error = if output_neurons.is_empty() {
-        0.0
-    } else {
-        let errors: Vec<f32> = output_neurons
-            .iter()
-            .map(|neuron| {
-                let records = get_records_or_error(records_provider.as_ref(), &neuron.uuid)?;
-                Ok(average_absolute_error_from_records(&records))
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        errors.into_iter().fold(0.0, f32::max)
-    };
-
-    // Use activation-based impact calculation for more accurate MIN/MAX/IF statistics
-    let impact_map = compute_impacts_with_activations(creature, records_provider.as_ref())?;
-
-    // Pre-compute synapse counts for O(1) lookup
-    let synapse_counts = SynapseCounts::new(creature);
-
-    // Issue #206: Build squash map for gradient flow analysis
-    let squash_map = build_squash_map(creature);
-
-    // Build neurons with base metrics
-    let mut neurons: Vec<RankedNeuron> = selectable
-        .par_iter()
-        .map(|neuron| -> Result<RankedNeuron> {
-            let records = get_records_or_error(records_provider.as_ref(), &neuron.uuid)?;
-            let raw_error = average_absolute_error_from_records(&records);
-            let structural_impact = *impact_map.get(&neuron.uuid).unwrap_or(&0.0);
-            let mean_activation = mean_absolute_activation_from_records(&records);
-            let activation_weighted_impact = structural_impact * mean_activation;
-
-            let total_error = if max_output_error > 0.0 {
-                raw_error.min(max_output_error)
-            } else {
-                raw_error
-            };
-
-            // Issue #206: Compute gradient flow stats for this neuron
-            let gradient_flow =
-                compute_gradient_flow_for_neuron(&neuron.uuid, &squash_map, &records);
-
-            // Issue #204: Compute activation frequency for focus neuron ranking
-            let activation_frequency = activation_frequency_from_records(&records);
-
-            Ok(RankedNeuron {
-                neuron_uuid: neuron.uuid.clone(),
-                total_error,
-                raw_error,
-                impact: structural_impact,
-                mean_activation,
-                activation_weighted_impact,
-                gradient_flow,
-                activation_frequency,
-            })
-        })
-        .collect::<Result<Vec<_>>>()?;
-
-    // Sort by weighted score with optional history factor, gradient flow factor, and frequency factor
-    // Issue #227: Incorporate historical success rate into ranking
-    // Issue #206: Incorporate gradient flow analysis into ranking
-    const IMPACT_EPSILON: f32 = 0.0001;
-    const IMPACT_GAMMA: f32 = 0.8;
-
-    neurons.sort_by(|a, b| {
-        let a_base = a.total_error * (a.impact + IMPACT_EPSILON).powf(IMPACT_GAMMA);
-        let b_base = b.total_error * (b.impact + IMPACT_EPSILON).powf(IMPACT_GAMMA);
-
-        // Issue #206: Apply gradient flow factor
-        let a_gradient_factor = compute_gradient_flow_factor(&a.gradient_flow);
-        let b_gradient_factor = compute_gradient_flow_factor(&b.gradient_flow);
-
-        // Issue #204: Apply activation frequency factor
-        let a_frequency_factor = compute_frequency_factor(a.activation_frequency);
-        let b_frequency_factor = compute_frequency_factor(b.activation_frequency);
-
-        let a_with_gradient = a_base * a_gradient_factor * a_frequency_factor;
-        let b_with_gradient = b_base * b_gradient_factor * b_frequency_factor;
-
-        // Apply history factor if available
-        // History factor is in [0, 1], where 0.5 is neutral
-        // We scale it so that:
-        // - 0.5 (neutral) → multiplier of 1.0 (no change)
-        // - 1.0 (perfect success) → multiplier of 1.5 (50% boost)
-        // - 0.0 (complete failure) → multiplier of 0.5 (50% penalty)
-        // Formula: multiplier = 0.5 + history_score
-        let (a_weighted, b_weighted) = if let Some(h) = history {
-            let a_history = h.bayesian_score_for(&a.neuron_uuid) as f32;
-            let b_history = h.bayesian_score_for(&b.neuron_uuid) as f32;
-            let a_multiplier = 0.5 + a_history;
-            let b_multiplier = 0.5 + b_history;
-            (
-                a_with_gradient * a_multiplier,
-                b_with_gradient * b_multiplier,
-            )
-        } else {
-            (a_with_gradient, b_with_gradient)
-        };
-
-        b_weighted
-            .total_cmp(&a_weighted)
-            .then_with(|| b.impact.total_cmp(&a.impact))
-            .then_with(|| a.neuron_uuid.cmp(&b.neuron_uuid))
-    });
-
-    // Identify removal candidates (same logic as rank_focus_neurons)
-    const DEFAULT_COST_OF_GROWTH: f32 = 1e-7;
-    let cost_of_growth_threshold = cost_of_growth.unwrap_or(DEFAULT_COST_OF_GROWTH);
-
-    let mut removal_candidates: Vec<RemovalCandidate> = neurons
-        .par_iter()
-        .filter_map(|n| {
-            let (incoming, outgoing) = synapse_counts.get(&n.neuron_uuid);
-            let savings = calculate_removal_savings(incoming, outgoing, cost_of_growth_threshold);
-
-            if savings <= n.activation_weighted_impact {
-                return None;
-            }
-
-            let expected_error_reduction = n.activation_weighted_impact;
-            let net_improvement = savings - n.activation_weighted_impact;
-
-            Some(RemovalCandidate {
-                neuron_uuid: n.neuron_uuid.clone(),
-                total_error: n.total_error,
-                impact: n.impact,
-                mean_activation: n.mean_activation,
-                activation_weighted_impact: n.activation_weighted_impact,
-                incoming_synapses: incoming,
-                outgoing_synapses: outgoing,
-                removal_savings: savings,
-                expected_error_reduction,
-                reason: format!(
-                    "Removal improves score: saves {:.2e} > impact {:.2e} (net +{:.2e}), {} synapses, costOfGrowth={:.2e}",
-                    savings,
-                    n.activation_weighted_impact,
-                    net_improvement,
-                    incoming + outgoing,
-                    cost_of_growth_threshold,
-                ),
-            })
-        })
-        .collect();
-
-    // Issue #414: High-error exploratory ablation DISABLED (see rank_focus_neurons for rationale)
-
-    // Sort removal candidates by net improvement
-    removal_candidates.sort_by(|a, b| {
-        let a_net = a.removal_savings - a.activation_weighted_impact;
-        let b_net = b.removal_savings - b.activation_weighted_impact;
-
-        b_net
-            .total_cmp(&a_net)
-            .then_with(|| {
-                a.activation_weighted_impact
-                    .total_cmp(&b.activation_weighted_impact)
-            })
-            .then_with(|| a.neuron_uuid.cmp(&b.neuron_uuid))
-    });
-
-    if let Some(limit) = max_results
-        && neurons.len() > limit
-    {
-        neurons.truncate(limit);
-    }
-
-    // Constant neuron removals (same as rank_focus_neurons)
-    let neuron_types: HashMap<&str, &str> = creature
-        .neurons
-        .iter()
-        .map(|n| (n.uuid.as_str(), n.neuron_type.as_str()))
-        .collect();
-
-    let neuron_biases: HashMap<&str, f32> = creature
-        .neurons
-        .iter()
-        .map(|n| (n.uuid.as_str(), n.bias))
-        .collect();
-
-    let outgoing_synapses: HashMap<&str, Vec<(&str, f32)>> =
-        creature
-            .synapses
-            .iter()
-            .fold(HashMap::new(), |mut map, syn| {
-                map.entry(syn.from_uuid.as_str())
-                    .or_default()
-                    .push((syn.to_uuid.as_str(), syn.weight));
-                map
-            });
-
-    let constant_neuron_removals: Vec<CoordinatedStructuralCandidateJson> = selectable
-        .par_iter()
-        .filter_map(|neuron| {
-            let neuron_type = neuron_types.get(neuron.uuid.as_str())?;
-            if *neuron_type != "hidden" {
-                return None;
-            }
-
-            let records = get_records_or_error(records_provider.as_ref(), &neuron.uuid).ok()?;
-            let (mean_activation, variance) = activation_mean_and_variance_from_records(&records);
-
-            if variance > CONSTANT_VARIANCE_THRESHOLD {
-                return None;
-            }
-
-            let outgoing = outgoing_synapses.get(neuron.uuid.as_str())?;
-            if outgoing.is_empty() {
-                return None;
-            }
-
-            let mut operations = Vec::with_capacity(outgoing.len() + 1);
-
-            for (to_uuid, weight) in outgoing {
-                let old_bias = neuron_biases.get(to_uuid).copied().unwrap_or(0.0);
-                let bias_adjustment = weight * mean_activation;
-                let new_bias = old_bias + bias_adjustment;
-
-                if new_bias.is_finite() {
-                    operations.push(CoordinatedStructuralOpJson::SetBias {
-                        neuron_uuid: to_uuid.to_string(),
-                        bias: new_bias,
-                    });
-                }
-            }
-
-            operations.push(CoordinatedStructuralOpJson::RemoveNeuron {
-                neuron_uuid: neuron.uuid.clone(),
-            });
-
-            let (incoming_count, outgoing_count) = synapse_counts.get(&neuron.uuid);
-            let removal_savings =
-                calculate_removal_savings(incoming_count, outgoing_count, cost_of_growth_threshold);
-
-            Some(CoordinatedStructuralCandidateJson {
-                operations,
-                expected_creature_score_gain: removal_savings,
-                comment: Some(format!(
-                    "Issue #306: Constant neuron removal with bias adjustments. \
-                     mean_activation={mean_activation:.6}, variance={variance:.2e}, \
-                     {} downstream neurons, savings={removal_savings:.2e}",
-                    outgoing.len()
-                )),
-            })
-        })
-        .collect();
-
-    Ok(RankFocusStats {
-        neurons,
-        removal_candidates,
-        constant_neuron_removals,
-        max_output_error,
-        processed_neurons: total_neurons,
-        total_neurons,
-        duration_ms: start.elapsed().as_millis(),
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use anyhow::anyhow;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    #[test]
-    fn lazy_provider_defers_loading_and_bounds_cache() -> Result<()> {
-        let loads = Arc::new(AtomicUsize::new(0));
-        let provider = LazyRecordProvider::with_loader_for_tests("unused.parquet", 2, {
-            let loads = Arc::clone(&loads);
-            Arc::new(move |_file, neuron_uuid| {
-                loads.fetch_add(1, Ordering::SeqCst);
-                Ok(vec![DiscoverRecord {
-                    obs_index: 0,
-                    neuron_uuid: neuron_uuid.to_string(),
-                    value: None,
-                    activation: 0.0,
-                    errors: vec![0.0],
-                }])
-            })
-        });
-
-        // No eager loads during initialisation
-        assert_eq!(0, loads.load(Ordering::SeqCst));
-
-        // First load hits the loader, subsequent load for same neuron is cached
-        provider.get("a")?.expect("records should be present");
-        assert_eq!(1, loads.load(Ordering::SeqCst));
-        provider.get("a")?.expect("records should be cached");
-        assert_eq!(1, loads.load(Ordering::SeqCst));
-
-        // Loading a second neuron increments once and cache remains bounded
-        provider.get("b")?.expect("records should be present");
-        assert_eq!(2, loads.load(Ordering::SeqCst));
-        assert!(provider.len() <= 2);
-        Ok(())
-    }
-
-    #[test]
-    fn lazy_provider_returns_loader_errors_with_context() {
-        let provider = LazyRecordProvider::with_loader_for_tests("failing.parquet", 2, {
-            Arc::new(|file, neuron_uuid| {
-                Err(anyhow!(
-                    "Simulated parquet read failure for {neuron_uuid} in {file}"
-                ))
-            })
-        });
-
-        let err = provider
-            .get("hidden-1")
-            .expect_err("loader error should surface");
-        let msg = format!("{err:#}");
-        assert!(
-            msg.contains("Simulated parquet read failure"),
-            "expected loader error, got: {msg}"
-        );
-        assert!(
-            msg.contains("hidden-1"),
-            "neuron context should be present: {msg}"
-        );
-        assert!(
-            msg.contains("failing.parquet"),
-            "file context should be present: {msg}"
-        );
-    }
-}
-
-// NOTE: Tests for focus module have been moved to tests/focus.rs
-// following the testing philosophy documented in README.md
-// (prefer tests/ over inline unit tests for tests that use public APIs).
